% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_utils.R
\name{as_sfn_data}
\alias{as_sfn_data}
\title{SfnData to sfn_data conversion}
\usage{
as_sfn_data(SfnData, parent_logger = "test")
}
\description{
This function converts an SfnData to sfn_data object
}
\details{
Things that this funcion do:
List of thing to do when passing from SfnData to sfn_data:
\enumerate{
\item pl_name in plant metadata must be set to character in all sites.
\item si_biome convert ot character (is factor)
\item tz of solarTIMESTAMP to UTC fixed
\item get rid of ascii characters in pl_sap_units and pl_sap_units_orig
\item leaf data has not pl_sap_units_orig and pl_sap_units does not match
due to bug in qc_sapw_conversion function
}
}
