% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control_functions.R
\name{coord_sign_test}
\alias{coord_sign_test}
\title{Coordinates sign test}
\usage{
coord_sign_test(data, maps_folder)
}
\arguments{
\item{data}{Data frame with data coming from \code{\link{check_coordinates}}
(with latitude, longitude, country and is_inside_country variables).}

\item{maps_folder}{Folder route where the maps are stored, by default the
working directory. It must be a character object and it must end
\bold{without} \code{/}.}
}
\value{
Same data frame provided, with a two new columns, \code{lat_changed}
  and \code{long_changed}, two logicals indicating if the coordinates are
  sign exchanged
}
\description{
\code{coord_sign_test} is an internal function to test if site coordinates
signs are interchanged. It's needed by \code{\link{fix_latlong_errors}}
function.
}
\details{
Country coordinates sign is established by this function and testing if
provided site coordinates are correct is made
}
\seealso{
Other Quality.Check.Functions: \code{\link{check_coordinates}},
  \code{\link{download_maps}}
}

