% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{df_copy_templates}
\alias{df_copy_templates}
\title{Save the Rmd templates and the running scripts to their corresponding folders}
\usage{
df_copy_templates(parent_logger = "test")
}
\value{
Logical indicating if all files have been copied or overwritten
}
\description{
Copy the Rmd templates for file transfer and quality check to Template folder and
the template for shiny web app to parent directory, and running scripts to parent
directory.
}
\seealso{
Other Data Flow: \code{\link{df_accepted_to_lvl1}},
  \code{\link{df_folder_structure}},
  \code{\link{df_get_data_folders}},
  \code{\link{df_get_status}},
  \code{\link{df_received_to_accepted}},
  \code{\link{df_report_folder_creation}},
  \code{\link{df_reset_data_status}},
  \code{\link{df_set_status}},
  \code{\link{df_start_status}}
}

