% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{df_flag_to_lvl2_app}
\alias{df_flag_to_lvl2_app}
\title{Interactively select the data sets ready to move to level 2}
\usage{
df_flag_to_lvl2_app(parent_logger = "test")
}
\value{
An interactive app to flag datasets as ready to move to level 2
}
\description{
Shiny app to flag data sets as ready to move to level 2
}
\seealso{
Other apps: \code{\link{out_app}}
}

