% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{df_flag_to_units_app}
\alias{df_flag_to_units_app}
\title{Interactively select the data sets ready for the units transformation}
\usage{
df_flag_to_units_app(parent_logger = "test")
}
\value{
An interactive app to flag datasets as ready to move to to units
  transformation
}
\description{
Shiny app to flag data sets as ready to move to units transformation
}
\seealso{
Other apps: \code{\link{df_flag_to_lvl2_app}},
  \code{\link{out_app}}, \code{\link{out_confirmation_app}}
}
