% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{df_folder_structure}
\alias{df_folder_structure}
\title{Initial project folder structure}
\usage{
df_folder_structure(parent_dir = ".", parent_logger = "test")
}
\arguments{
\item{parent_dir}{Character vector indicating the parent directory where
the structure must be built. Default to current directory.}
}
\value{
Nothing, dirs are created if they do not exist, or a warning is raised
if they already exist.
}
\description{
Create the folder tree skeleton for SAPFLUXNET project
}
\details{
This functions only need to be called once, as it creates the needed folder
structure to recreate SAPFLUXNET project in any environment.
}
\seealso{
Other Data Flow: 
\code{\link{df_accepted_to_lvl1}()},
\code{\link{df_copy_templates}()},
\code{\link{df_get_data_folders}()},
\code{\link{df_get_status}()},
\code{\link{df_lvl1_to_lvl2}()},
\code{\link{df_lvl2_folder_structure}()},
\code{\link{df_read_SfnData}()},
\code{\link{df_received_to_accepted}()},
\code{\link{df_rem_to_units}()},
\code{\link{df_report_folder_creation}()},
\code{\link{df_reset_data_status}()},
\code{\link{df_set_status}()},
\code{\link{df_start_status}()},
\code{\link{df_warn_to_rem}()},
\code{\link{df_whos_ready_to}()},
\code{\link{df_write_SfnData}()},
\code{\link{sfn_data_constructor}()}
}
\concept{Data Flow}
