% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{df_get_data_folders}
\alias{df_get_data_folders}
\title{Get the folders to feed \code{\link{qc_start_process}}}
\usage{
df_get_data_folders(parent_logger = "test")
}
\value{
a character vector with the folders route
}
\description{
Get the folders to be able to start QC process
}
\details{
No parameters (except for parent_logger) are needed, as the functions simply
collects the folders present in \code{Data}.
}
\seealso{
Other Data Flow: \code{\link{df_accepted_to_lvl1}},
  \code{\link{df_copy_templates}},
  \code{\link{df_folder_structure}},
  \code{\link{df_get_status}},
  \code{\link{df_lvl2_folder_structure}},
  \code{\link{df_received_to_accepted}},
  \code{\link{df_report_folder_creation}},
  \code{\link{df_reset_data_status}},
  \code{\link{df_set_status}},
  \code{\link{df_start_status}},
  \code{\link{df_who_ready_to_lvl2}},
  \code{\link{sfn_data_constructor}}
}

