% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{df_lvl1_to_lvl2}
\alias{df_lvl1_to_lvl2}
\title{Function to pass from level 1 to level 2}
\usage{
df_lvl1_to_lvl2(parent_logger = "test")
}
\value{
Nothing, all the process is internal
}
\description{
LVL1 to LVL2 transfer
}
\details{
This function is in charge of check for sites ready to pass to Level 2, create
the needed folder structure, flag for outliers warnings and saving the final
SfnData objects in the corresponding folder.
}
\seealso{
Other Data Flow: \code{\link{df_accepted_to_lvl1}},
  \code{\link{df_copy_templates}},
  \code{\link{df_folder_structure}},
  \code{\link{df_get_data_folders}},
  \code{\link{df_get_status}},
  \code{\link{df_lvl2_folder_structure}},
  \code{\link{df_read_SfnData}},
  \code{\link{df_received_to_accepted}},
  \code{\link{df_report_folder_creation}},
  \code{\link{df_reset_data_status}},
  \code{\link{df_set_status}},
  \code{\link{df_start_status}},
  \code{\link{df_whos_ready_to}},
  \code{\link{df_write_SfnData}},
  \code{\link{sfn_data_constructor}}
}
