% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{df_lvl2_folder_structure}
\alias{df_lvl2_folder_structure}
\title{Build the folder structure of level 2}
\usage{
df_lvl2_folder_structure(si_code, parent_logger = "test")
}
\arguments{
\item{si_code}{Site code as character}
}
\value{
Nothing, if dir is created ok silent exit, if not it throws an error
}
\description{
Take a site code and build the level 2 tree folder
}
\seealso{
Other Data Flow: \code{\link{df_accepted_to_lvl1}},
  \code{\link{df_copy_templates}},
  \code{\link{df_folder_structure}},
  \code{\link{df_get_data_folders}},
  \code{\link{df_get_status}},
  \code{\link{df_lvl1_to_lvl2}},
  \code{\link{df_read_SfnData}},
  \code{\link{df_received_to_accepted}},
  \code{\link{df_report_folder_creation}},
  \code{\link{df_reset_data_status}},
  \code{\link{df_set_status}},
  \code{\link{df_start_status}},
  \code{\link{df_whos_ready_to}},
  \code{\link{df_write_SfnData}},
  \code{\link{sfn_data_constructor}}
}
