% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{df_reset_data_status}
\alias{df_reset_data_status}
\title{Reset the data folder and status file for the si_code given.}
\usage{
df_reset_data_status(si_code, level = "all", parent_logger = "test")
}
\arguments{
\item{si_code}{Character vector indicating the site code to reset}

\item{level}{Character string indicating which level must be resetted. Valid
values are "LVL1", "LVL2" and "all"}
}
\value{
Nothing
}
\description{
After manual changes, update the satus file to indicate that QC is needed.
Also rename Accepted and Lvl1 data to avoid conflicts with manually
changed data.
}
\details{
A fast way of reset any site data folder when needed, usually after manual
changes of the files.
\code{level} parameter allows for selecting to which level the status must
be resetted. "LVL1" indicates that only LVL1 data and status will be
resetted. "LVL2" resets only LVL2 (warn, rem, units) data and status and
"all" resets all data and status to the initial state for the site.
}
\seealso{
Other Data Flow: \code{\link{df_accepted_to_lvl1}},
  \code{\link{df_copy_templates}},
  \code{\link{df_folder_structure}},
  \code{\link{df_get_data_folders}},
  \code{\link{df_get_status}},
  \code{\link{df_lvl1_to_lvl2}},
  \code{\link{df_lvl2_folder_structure}},
  \code{\link{df_read_SfnData}},
  \code{\link{df_received_to_accepted}},
  \code{\link{df_report_folder_creation}},
  \code{\link{df_set_status}},
  \code{\link{df_start_status}},
  \code{\link{df_warn_to_rem}},
  \code{\link{df_whos_ready_to}},
  \code{\link{df_write_SfnData}},
  \code{\link{sfn_data_constructor}}
}
