% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{df_start_status}
\alias{df_start_status}
\title{Initialise an empty status file}
\usage{
df_start_status(si_code, parent_logger = "test")
}
\arguments{
\item{si_code}{Character vector indicating the site code}
}
\value{
Invisible TRUE if no errors were encountered, invisible FALSE if
there was errors. Also, status file is created in the corresponding folder.
}
\description{
Initialise an empty status file in yaml format, using the yaml package
}
\details{
Before creating an empty file, \code{df_start_status} checks if an status
file already exists, in order to avoid accidental rewriting of the file.
}
\seealso{
Other Data Flow: 
\code{\link{df_accepted_to_lvl1}()},
\code{\link{df_copy_templates}()},
\code{\link{df_folder_structure}()},
\code{\link{df_get_data_folders}()},
\code{\link{df_get_status}()},
\code{\link{df_lvl1_to_lvl2}()},
\code{\link{df_lvl2_folder_structure}()},
\code{\link{df_read_SfnData}()},
\code{\link{df_received_to_accepted}()},
\code{\link{df_rem_to_units}()},
\code{\link{df_report_folder_creation}()},
\code{\link{df_reset_data_status}()},
\code{\link{df_set_status}()},
\code{\link{df_warn_to_rem}()},
\code{\link{df_whos_ready_to}()},
\code{\link{df_write_SfnData}()},
\code{\link{sfn_data_constructor}()}
}
\concept{Data Flow}
