% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{df_warn_to_rem}
\alias{df_warn_to_rem}
\title{Function to pass from out_warn to out_rem}
\usage{
df_warn_to_rem(parent_logger = "test")
}
\value{
Nothing, all the process is internal
}
\description{
WARN to REM transfer
}
\details{
This function is in charge of check for sites ready to pass to out_rem,
remove the outliers indicated in the accesory files and saving the final
SfnData objects in the corresponding folder (out_rem).
}
\seealso{
Other Data Flow: \code{\link{df_accepted_to_lvl1}},
  \code{\link{df_copy_templates}},
  \code{\link{df_folder_structure}},
  \code{\link{df_get_data_folders}},
  \code{\link{df_get_status}},
  \code{\link{df_lvl1_to_lvl2}},
  \code{\link{df_lvl2_folder_structure}},
  \code{\link{df_read_SfnData}},
  \code{\link{df_received_to_accepted}},
  \code{\link{df_rem_to_units}},
  \code{\link{df_report_folder_creation}},
  \code{\link{df_reset_data_status}},
  \code{\link{df_set_status}},
  \code{\link{df_start_status}},
  \code{\link{df_whos_ready_to}},
  \code{\link{df_write_SfnData}},
  \code{\link{sfn_data_constructor}}
}
