% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{df_who_ready_to_lvl2}
\alias{df_who_ready_to_lvl2}
\title{Who is ready for level 2?}
\usage{
df_who_ready_to_lvl2(filter = c("all", "ready", "freeze", "done"),
  parent_logger = "test")
}
\arguments{
\item{filter}{character vector indicating by which TO_LVL2 status results must
been filtered. Accepted values are "all" (default), "ready", "freeze" or "done"
(see details)}
}
\value{
A list with length equal to the number of sites containing the
  TO_LVL2 flag of the status files.
}
\description{
Check the site status files to list who is ready to move to level 2
}
\details{
\code{filter} parameter indicates if the results must be filtered by the
  corresponding status:

\itemize{
    \item{\code{all} retrieves all the statuses}
    \item{\code{ready} retrieves only those sites marked to pass to level 2}
    \item{\code{freeze} retrieves only those sites freezed in level 1 yet}
    \item{\code{done} retrieves only those sites already passed to level 2}
  }
}
\seealso{
Other Data Flow: \code{\link{df_accepted_to_lvl1}},
  \code{\link{df_copy_templates}},
  \code{\link{df_folder_structure}},
  \code{\link{df_get_data_folders}},
  \code{\link{df_get_status}},
  \code{\link{df_lvl2_folder_structure}},
  \code{\link{df_received_to_accepted}},
  \code{\link{df_report_folder_creation}},
  \code{\link{df_reset_data_status}},
  \code{\link{df_set_status}},
  \code{\link{df_start_status}},
  \code{\link{sfn_data_constructor}}
}

