% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{df_write_SfnData}
\alias{df_write_SfnData}
\title{Write SfnData}
\usage{
df_write_SfnData(SfnData, level = c("Lvl_1", "Lvl_2", "out_warn", "out_rem",
  "unit_trans"), parent_logger = "test")
}
\arguments{
\item{SfnData}{SfnData object}

\item{level}{Level to write to as a character string}
}
\value{
Nothing, the desired site data is saved.
}
\description{
Accesory function to write an SfnData object to a fixed location
}
\details{
Given a site code and a level description, \code{df_write_SfnData} will save
the selected SfnData object in the selected location
}
\seealso{
Other Data Flow: \code{\link{df_accepted_to_lvl1}},
  \code{\link{df_copy_templates}},
  \code{\link{df_folder_structure}},
  \code{\link{df_get_data_folders}},
  \code{\link{df_get_status}},
  \code{\link{df_lvl1_to_lvl2}},
  \code{\link{df_lvl2_folder_structure}},
  \code{\link{df_read_SfnData}},
  \code{\link{df_received_to_accepted}},
  \code{\link{df_rem_to_units}},
  \code{\link{df_report_folder_creation}},
  \code{\link{df_reset_data_status}},
  \code{\link{df_set_status}},
  \code{\link{df_start_status}},
  \code{\link{df_warn_to_rem}},
  \code{\link{df_whos_ready_to}},
  \code{\link{sfn_data_constructor}}
}
