% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_data_loading_functions.R
\name{dl_data}
\alias{dl_data}
\title{Loading data from xls/xlsx ann csv files}
\usage{
dl_data(
  file_name,
  sheet_name,
  long = FALSE,
  n = 1000,
  na = "",
  parent_logger = "test"
)
}
\arguments{
\item{file_name}{Character vector indicating the name of the file
containing the data.}

\item{sheet_name}{Character vector indicating the name of the sheet to be
loaded. It must be one of \code{sapflow_hd} or \code{environmental_hd}. It
is only used if origin file is an xlsx file}

\item{long}{Logical indicating if returned data must be in \code{long} format}

\item{n}{Numeric indicating the number of rows used to guess the decimal
character used in csv files. Only used when data file is a csv.}

\item{na}{Character string with the symbol codified as missing value.
Default to blank.}
}
\value{
\code{dl_data} returns sapflow or environmental data in wide format,
ready to pipe it in the quality checks for raw data. If \code{long = TRUE}
data is returned in long format, ready for plotting.
}
\description{
Detecting file type and load data from files.
}
\details{
This function make use of dplyr, tidyr and readxl packages in order to
retrieve and format the data. Also, in the case of csv files it uses
data.table package.
}
\seealso{
Other Data Loading Functions: 
\code{\link{dl_data_col_classes}()},
\code{\link{dl_dec_char_detect}()},
\code{\link{dl_get_si_code}()},
\code{\link{dl_metadata}()},
\code{\link{dl_na_char_generator}()},
\code{\link{qc_as_timestamp}()},
\code{\link{remove_dupcols}()}
}
\concept{Data Loading Functions}
