% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_data_loading_functions.R
\name{dl_data_col_classes}
\alias{dl_data_col_classes}
\title{Check data columns classes and set to numeric if needed}
\usage{
dl_data_col_classes(data, parent_logger = "test")
}
\arguments{
\item{data}{Data frame with the sapflow or environmental data to check}
}
\value{
A data frame exactly as the data provided with the variables
changed to numeric if needed
}
\description{
Checking and setting data column classes to numeric
}
\details{
Loading from csv's with fread can be problematic if some characters are
spreaded here and there in the data columns. This function checks for this
problem and tries to fix it
}
\seealso{
Other Data Loading Functions: 
\code{\link{dl_data}()},
\code{\link{dl_dec_char_detect}()},
\code{\link{dl_get_si_code}()},
\code{\link{dl_metadata}()},
\code{\link{dl_na_char_generator}()},
\code{\link{qc_as_timestamp}()},
\code{\link{remove_dupcols}()}
}
\concept{Data Loading Functions}
