% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control_functions.R
\name{download_maps}
\alias{download_maps}
\title{Download maps for countries included in the database}
\usage{
download_maps(data, folder)
}
\arguments{
\item{data}{Dataframe where the countries ISO codes are. Must contain a
variable called \code{country} where the ISO code resides.}

\item{folder}{Folder where the maps are stored or where they will be stored,
by default the working directory. It must be a character.}
}
\value{
Maps are downloaded (if needed) and a summary is returned indicating
  number of maps downlaoded and number of maps present in the map folder.
}
\description{
\code{download_maps} fetch maps from \url{http://www.gadm.org/}.
}
\details{
This function fetch maps from \url{http://www.gadm.org/} if the map it's not
already present in the maps folder (by default, the working directory).
}

