% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_functions.R
\name{log_sapfluxnet_action}
\alias{log_sapfluxnet_action}
\title{Logging handler action}
\usage{
log_sapfluxnet_action(msg, handler, ...)
}
\arguments{
\item{msg}{The final formatted message to be output}

\item{handler}{The handler owning this action}
}
\description{
Custom logging handler action for SAPFLUXNET Project
}
\seealso{
Other Logging Functions: \code{\link{log_sapfluxnet_format}},
  \code{\link{log_sapfluxnet_setup}}
}

