% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_functions.R
\name{log_sapfluxnet_format}
\alias{log_sapfluxnet_format}
\title{Logs formatter}
\usage{
log_sapfluxnet_format(record)
}
\arguments{
\item{record}{log-record object from log action}
}
\value{
Formatted message for log-record
}
\description{
Custom log formatter for SAPFLUXNET Project
}
\seealso{
Other Logging Functions: \code{\link{log_sapfluxnet_action}},
  \code{\link{log_sapfluxnet_setup}}
}
