% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_functions.R
\name{log_sapfluxnet_setup}
\alias{log_sapfluxnet_setup}
\title{Setup handler}
\usage{
log_sapfluxnet_setup(file_name, logger)
}
\arguments{
\item{file_name}{Name of the file in which the log will be saved}

\item{logger}{Name of the logger}
}
\description{
Setting up the SAPFLUXNET Project log handler
}
\seealso{
Other Logging Functions: \code{\link{log_sapfluxnet_action}},
  \code{\link{log_sapfluxnet_format}}
}

