% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_outliers_app.R
\name{out_app}
\alias{out_app}
\title{outliers_app function}
\usage{
out_app(parent_logger = "test")
}
\value{
An interactive app to manage outliers
}
\description{
Interactive app to flag outliers, out_of_range data and visually detected
data flaws.
}
\details{
This function launches a shiny app to visually manage flagging outliers,
out_of_range values and visual data flaws. Data values selected after
visual exploration of the data presented in the app are coded in files
saved in the \code{Data/si_code/Lvl_2/lvl_2_out_warn} folder. These files
are used later to substitute outliers, remove out of range values and
remove manual values in the \code{\link{df_warn_to_remove}} function.
}
\seealso{
Other apps: \code{\link{df_flag_to_lvl2_app}},
  \code{\link{df_flag_to_units_app}},
  \code{\link{out_confirmation_app}}
}
