% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_outliers_app.R
\name{out_confirmation_app}
\alias{out_confirmation_app}
\title{outlier confirmation app}
\usage{
out_confirmation_app(parent_logger = "test")
}
\value{
An interactive app to manage outliers
}
\description{
Interactive app to confirm the data after removing outliers, out of range
values and other modifications
}
\seealso{
Other apps: \code{\link{df_flag_to_lvl2_app}},
  \code{\link{df_flag_to_units_app}}, \code{\link{out_app}}
}
