% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_outliers.R
\name{out_remove}
\alias{out_remove}
\title{Outliers QC}
\usage{
out_remove(sfn_data, k = 25L, t0 = 10L, method = "hampel",
  reverse = TRUE, substitute = FALSE, parent_logger = "test")
}
\arguments{
\item{sfn_data}{SfnData object with the site data and metadata}

\item{k}{window semi-value (integer) for substitution value calculation. See
\code{\link{out_hampel_filter}} for details.}

\item{t0}{Integer value indicating the number of standard deviations for
outlier detection threshold.}

\item{method}{Character vector indicating the method to use in the outlier
estimation: \code{hampel} (default), \code{tukey} and \code{quantile}.}

\item{reverse}{Logical indicating if in case of k is bigger than the data
length (i.e in the extremes of the vector) a reverse replication must be
done. Default to TRUE.}

\item{substitute}{Logical indicating if the outlier subsitution must be made.
Default to FALSE.}
}
\value{
a SfnData object as the one provided with the oulier values
  substituted and the flags slot updated.
}
\description{
Outliers detection, substitution and annotation for SfnData objects
}
\details{
Outliers for sap flow data and environmental data are located and substituted
by the selected method, using the \code{\link{out_hampel_filter}} function.
}
\seealso{
Other outliers: \code{\link{out_hampel_filter}},
  \code{\link{outlier_subs}}
}

