% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_outliers.R
\name{outlier_subs}
\alias{outlier_subs}
\alias{out_tukeyline}
\alias{out_medianreg}
\alias{out_median}
\title{Outliers substitution}
\usage{
out_tukeyline(y, k = 5L, parent_logger = "test")

out_medianreg(y, k = 5L, parent_logger = "test")

out_median(y, k = 5L, parent_logger = "test")
}
\arguments{
\item{y}{vector of values for outlier substitution}

\item{k}{window semi-value (integer) for substitution value calculation. See
details.}
}
\value{
a vector of the same lengh of y with the outlier values substituted
  by the calculated value
}
\description{
Outlier substitution by different methods
}
\details{
\code{k} is the semi-value of the desired window for substitute value
calculation. The window is formed by the range \code{y[i] - k : y[i] + k}.
}
\section{Functions}{
\itemize{
\item \code{out_tukeyline}: 

\item \code{out_medianreg}: 

\item \code{out_median}: 
}}

\section{out_tukeyline}{

  \code{out_tukeyline} performs a robust fitting with the \code{line} function
}

\section{out_medianreg}{

  \code{out_medianreg} performs a regression of medians with the quantreg package
}

\section{out_median}{

  \code{out_median} performs a classic median outlier detection
}

\seealso{
Other outliers: \code{\link{out_hampel_filter}},
  \code{\link{out_remove}}
}
