% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SfnData_methods.R
\docType{methods}
\name{plot,SfnData,missing-method}
\alias{plot,SfnData,missing-method}
\title{plot SfnData method}
\usage{
\S4method{plot}{SfnData,missing}(x, type = c("sapf", "env", "ta", "rh", "vpd",
  "ppfd_in", "netrad", "sw_in", "ext_rad", "ws", "precip", "swc_shallow",
  "swc_deep"), solar = FALSE)
}
\arguments{
\item{type}{what to plot}

\item{solar}{use solarTIMESTAMP?}

\item{object}{SfnData object}
}
\description{
plot SfnData method
}
