% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_unit_transformations.R
\name{qc_cm_cm_h}
\alias{qc_cm_cm_h}
\alias{qc_cm_m_s}
\alias{qc_dm_dm_h}
\alias{qc_dm_dm_s}
\alias{qc_mm_mm_s}
\alias{qc_g_m_s}
\alias{qc_kg_m_h}
\alias{qc_kg_m_s}
\alias{qc_cm_h}
\alias{qc_dm_h}
\alias{qc_cm_s}
\alias{qc_g_h}
\alias{qc_kg_h}
\title{Unit conversion}
\usage{
qc_cm_cm_h(x, sapw_area, leaf_area, output_units, parent_logger = "test")

qc_cm_m_s(x, sapw_area, leaf_area, output_units, parent_logger = "test")

qc_dm_dm_h(x, sapw_area, leaf_area, output_units, parent_logger = "test")

qc_dm_dm_s(x, sapw_area, leaf_area, output_units, parent_logger = "test")

qc_mm_mm_s(x, sapw_area, leaf_area, output_units, parent_logger = "test")

qc_g_m_s(x, sapw_area, leaf_area, output_units, parent_logger = "test")

qc_kg_m_h(x, sapw_area, leaf_area, output_units, parent_logger = "test")

qc_kg_m_s(x, sapw_area, leaf_area, output_units, parent_logger = "test")

qc_cm_h(x, sapw_area, leaf_area, output_units, parent_logger = "test")

qc_dm_h(x, sapw_area, leaf_area, output_units, parent_logger = "test")

qc_cm_s(x, sapw_area, leaf_area, output_units, parent_logger = "test")

qc_g_h(x, sapw_area, leaf_area, output_units, parent_logger = "test")

qc_kg_h(x, sapw_area, leaf_area, output_units, parent_logger = "test")
}
\arguments{
\item{x}{Numeric value in which the conversion must be done}

\item{sapw_area}{Numeric value with the plant sapwood area in cm²}

\item{leaf_area}{Numeric value with the plant leaf area in m²}

\item{output_units}{Character vector indicating the kind of output units.
Allowed values are \code{'plant'}, \code{'sapwood'} and \code{'leaf'}.}
}
\value{
A numeric value resulting from unit conversion
}
\description{
Conversion of sap flow units to sapwood level
(cm³·cm⁻²·h⁻¹), plant level (cm³·h⁻¹) or leaf area level (cm³·cm⁻²·h⁻¹)
}
\section{Functions}{
\itemize{
\item \code{qc_cm_m_s()}: Conversion of sap flow units from cm³·m⁻²·h⁻¹

\item \code{qc_dm_dm_h()}: Conversion of sap flow units from dm³·dm⁻²·h⁻¹

\item \code{qc_dm_dm_s()}: Conversion of sap flow units from dm³·dm⁻²·s⁻¹

\item \code{qc_mm_mm_s()}: Conversion of sap flow units from mm³·mm⁻²·s⁻¹

\item \code{qc_g_m_s()}: Conversion of sap flow units from g·m⁻²·s⁻¹

\item \code{qc_kg_m_h()}: Conversion of sap flow units from kg·m⁻²·h⁻¹

\item \code{qc_kg_m_s()}: Conversion of sap flow units from kg·m⁻²·s⁻¹

\item \code{qc_cm_h()}: Conversion of sap flow units from cm³·h⁻¹

\item \code{qc_dm_h()}: Conversion of sap flow units from dm³·h⁻¹

\item \code{qc_cm_s()}: Conversion of sap flow units from cm³·s⁻¹

\item \code{qc_g_h()}: Conversion of sap flow units from g·h⁻¹

\item \code{qc_kg_h()}: Conversion of sap flow units from kg·h⁻¹

}}
\seealso{
Other Unit conversion: 
\code{\link{qc_rh}()},
\code{\link{qc_transf_list}()},
\code{\link{qc_transformation_vars}()},
\code{\link{qc_units_process}()},
\code{\link{qc_vpd}()}
}
\concept{Unit conversion}
