% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_gaps_functions.R
\name{qc_mind_the_gap_deprecated}
\alias{qc_mind_the_gap_deprecated}
\title{Gaps info}
\usage{
qc_mind_the_gap_deprecated(data, trim = FALSE, parent_logger = "test")
}
\arguments{
\item{data}{Data frame containing the data in which obtain info about gaps.
It must have a TIMESTAMP variable}

\item{trim}{Logial indicating if starting and ending gaps must be included.
Default to FALSE.}
}
\value{
A data frame with info about gaps:
\itemize{
\item gap_start: Gap start TIMESTAMP
\item gap_end: Gap end TIMESTAMP
\item gap_interval: Gap interval in minutes
\item gap_coverage: Gap coverage in percentage
\item timestamp_start: Absolute timestamp start
\item timestamp_end: Absolute timestamp end
}
}
\description{
Information about gaps
}
\details{
Gaps are considered from the last TIMESTAMP with data to the next TIMESTAMP
with data. In this way, a gap of only one TIMESTAMP has a duration equal to
the timestep of the TIMESTAMP. The only exception to this is if the gap
includes the start or the end of the TIMESTAMP. In this case the gap start
or end in the first or the last TIMESTAMP
}
\seealso{
Other Quality Checks Functions: 
\code{\link{create_dic}()},
\code{\link{qc_coordinates}()},
\code{\link{qc_data_results_table}()},
\code{\link{qc_env_ranges}()},
\code{\link{qc_env_vars_presence}()},
\code{\link{qc_ext_radiation}()},
\code{\link{qc_factor_values}()},
\code{\link{qc_fix_timestamp}()},
\code{\link{qc_get_biome}()},
\code{\link{qc_get_biomes_spdf}()},
\code{\link{qc_get_sapw_md}()},
\code{\link{qc_get_timestep}()},
\code{\link{qc_get_timezone}()},
\code{\link{qc_is_timestamp}()},
\code{\link{qc_md_cols}()},
\code{\link{qc_md_results_table}()},
\code{\link{qc_mind_the_gap}()},
\code{\link{qc_mind_the_gap_eff}()},
\code{\link{qc_out_hampel_filter}()},
\code{\link{qc_out_of_range}()},
\code{\link{qc_out_remove}()},
\code{\link{qc_outliers_process}()},
\code{\link{qc_outliers_subs}},
\code{\link{qc_pl_treatments}()},
\code{\link{qc_rad_conversion}()},
\code{\link{qc_range_dic}()},
\code{\link{qc_sapf_ranges}()},
\code{\link{qc_sapw_area_calculator}()},
\code{\link{qc_sapw_conversion}()},
\code{\link{qc_set_timezone}()},
\code{\link{qc_soil_texture}()},
\code{\link{qc_species_names}()},
\code{\link{qc_species_names_info}()},
\code{\link{qc_species_verification}()},
\code{\link{qc_start_process}()},
\code{\link{qc_swc_check}()},
\code{\link{qc_swc_fix}()},
\code{\link{qc_time_interval}()},
\code{\link{qc_timestamp_concordance}()},
\code{\link{qc_timestamp_errors}()},
\code{\link{qc_timestamp_nas}()}
}
\concept{Quality Checks Functions}
