% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_unit_transformations.R
\name{qc_pl_units_helper}
\alias{qc_pl_units_helper}
\title{Helper function for unit conversion}
\usage{
qc_pl_units_helper(pl_data)
}
\arguments{
\item{pl_data}{Data frame containing the plant metadata}
}
\value{
A data frame with extracted plant metadata variables as columns and
  individual plants as rows.
}
\description{
\code{qc_pl_units_helper} allows to extract all necessary variables from plant
metadata in order to be able to convert sap flow units.
}
\seealso{
Other Quality Checks Functions: \code{\link{create_dic}},
  \code{\link{qc_env_vars_presence}},
  \code{\link{qc_factor_values}},
  \code{\link{qc_is_timestamp}}, \code{\link{qc_md_cols}},
  \code{\link{qc_pl_treatments}},
  \code{\link{qc_species_verification}},
  \code{\link{qc_timestamp_errors}}
}

