% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_gaps_functions.R
\name{qc_plot_the_gap}
\alias{qc_plot_the_gap}
\title{Plotting an histogram for gaps intervals/gaps coverage}
\usage{
qc_plot_the_gap(gaps_info, type = "gap_interval", binwidth = NULL,
  parent_logger = "test")
}
\arguments{
\item{gaps_info}{Data frame as obtained from \code{\link{qc_mind_the_gap}}}

\item{type}{Character indicating what to represent, \code{gap_interval} or
\code{gap_coverage}}

\item{binwidth}{Bin width as stated in geom_histogram, default to NULL to
use the geom_histrogram default. Change it if more or less resolution is
needed}
}
\value{
a ggplot object with the basic histogram, no themes added.
}
\description{
Wrapper for ggplot to plot an histogram of gaps info
}
\details{
This function is a simple wrapper for ggplot + geom_histogram. It produces
a ggplot object that can be modified adding layers, like any other ggplot
object.
}
\seealso{
Other Quality Checks Functions: \code{\link{create_dic}},
  \code{\link{qc_coordinates}},
  \code{\link{qc_data_results_table}},
  \code{\link{qc_env_vars_presence}},
  \code{\link{qc_factor_values}},
  \code{\link{qc_fix_timestamp}},
  \code{\link{qc_get_sapw_md}},
  \code{\link{qc_get_timestep}},
  \code{\link{qc_is_timestamp}}, \code{\link{qc_md_cols}},
  \code{\link{qc_md_results_table}},
  \code{\link{qc_mind_the_gap}},
  \code{\link{qc_pl_treatments}},
  \code{\link{qc_sapw_area_calculator}},
  \code{\link{qc_sapw_conversion}},
  \code{\link{qc_species_names_info}},
  \code{\link{qc_species_names}},
  \code{\link{qc_species_verification}},
  \code{\link{qc_time_interval}},
  \code{\link{qc_timestamp_concordance}},
  \code{\link{qc_timestamp_errors}}
}

