% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_unit_transformations.R
\name{qc_rad_conversion}
\alias{qc_rad_conversion}
\title{Radiation units transformation}
\usage{
qc_rad_conversion(data, parent_logger = "test")
}
\arguments{
\item{data}{Data frame containing the environmental measurements}
}
\description{
Function to transform between radiation units
}
\details{
Radiation accepted measures can be of two kinds, \code{incoming photosynthetic photon flux density}
and \code{shortwave incoming radiation}. Unit conversion must be done to allow data
integration and analysis.
This function converts between:
\describe{
\item{\bold{Incoming photosynthetic photon flux density}}{
In this case, units returned are \eqn{{\mu}mol·m⁻²·s⁻¹}
  }
  \item{\bold{Shortwave incoming radiation}}{
In this case, units returned are \eqn{W·m⁻²}
  }
}
}
\section{Incoming photosynthetic photon flux density}{

Direct transformation is made from \emph{shortwave incoming radiation}.
}

\section{Shortwave incoming radiation}{

Direct transformation is made from \emph{incoming photosynthetic photon flux density}.
}

\seealso{
Other Quality Checks Functions: 
\code{\link{create_dic}()},
\code{\link{qc_coordinates}()},
\code{\link{qc_data_results_table}()},
\code{\link{qc_env_ranges}()},
\code{\link{qc_env_vars_presence}()},
\code{\link{qc_ext_radiation}()},
\code{\link{qc_factor_values}()},
\code{\link{qc_fix_timestamp}()},
\code{\link{qc_get_biome}()},
\code{\link{qc_get_biomes_spdf}()},
\code{\link{qc_get_sapw_md}()},
\code{\link{qc_get_timestep}()},
\code{\link{qc_get_timezone}()},
\code{\link{qc_is_timestamp}()},
\code{\link{qc_md_cols}()},
\code{\link{qc_md_results_table}()},
\code{\link{qc_mind_the_gap}()},
\code{\link{qc_mind_the_gap_eff}()},
\code{\link{qc_out_hampel_filter}()},
\code{\link{qc_out_of_range}()},
\code{\link{qc_out_remove}()},
\code{\link{qc_outliers_process}()},
\code{\link{qc_outliers_subs}},
\code{\link{qc_pl_treatments}()},
\code{\link{qc_range_dic}()},
\code{\link{qc_sapf_ranges}()},
\code{\link{qc_sapw_area_calculator}()},
\code{\link{qc_sapw_conversion}()},
\code{\link{qc_set_timezone}()},
\code{\link{qc_soil_texture}()},
\code{\link{qc_species_names}()},
\code{\link{qc_species_names_info}()},
\code{\link{qc_species_verification}()},
\code{\link{qc_start_process}()},
\code{\link{qc_swc_check}()},
\code{\link{qc_swc_fix}()},
\code{\link{qc_time_interval}()},
\code{\link{qc_timestamp_concordance}()},
\code{\link{qc_timestamp_errors}()},
\code{\link{qc_timestamp_nas}()}
}
\concept{Quality Checks Functions}
