% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_numerical_metadata_functions.R
\name{qc_sapf_range_check}
\alias{qc_sapf_range_check}
\title{Suitable Range for sapflow values}
\usage{
qc_sapf_range_check(data, parent_logger = "test")
}
\arguments{
\item{data}{Data frame with the values of sapflow measures for each plant
(variable).}
}
\value{
A data frame with summarising the values outside of the suitable
  range.
}
\description{
Suitability check for sapflow values.
}
\details{
Ranges for sapflow measures at sapwood level are obtained from literature,
being the most extreme values -10 and 533 cm3cm-2h-1 (Manzoni et al 2013 and
other sources). This range is only informative, but values outside of the
range may indicate some error in the data.
}
\seealso{
Other Quality Checks Functions: \code{\link{create_dic}},
  \code{\link{qc_coordinates}},
  \code{\link{qc_data_results_table}},
  \code{\link{qc_env_vars_presence}},
  \code{\link{qc_factor_values}},
  \code{\link{qc_fix_timestamp}},
  \code{\link{qc_get_sapw_md}},
  \code{\link{qc_get_timestep}},
  \code{\link{qc_get_timezone}},
  \code{\link{qc_is_timestamp}}, \code{\link{qc_md_cols}},
  \code{\link{qc_md_results_table}},
  \code{\link{qc_mind_the_gap_eff}},
  \code{\link{qc_mind_the_gap}},
  \code{\link{qc_pl_treatments}},
  \code{\link{qc_sapw_area_calculator}},
  \code{\link{qc_sapw_conversion}},
  \code{\link{qc_set_timezone}},
  \code{\link{qc_solar_timestamp}},
  \code{\link{qc_species_names_info}},
  \code{\link{qc_species_names}},
  \code{\link{qc_species_verification}},
  \code{\link{qc_start_process}},
  \code{\link{qc_time_interval}},
  \code{\link{qc_timestamp_concordance}},
  \code{\link{qc_timestamp_errors}},
  \code{\link{qc_timestamp_nas}}
}

