% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_md_general_checks.R
\name{qc_site_col}
\alias{qc_site_col}
\title{Site metadata column check}
\usage{
qc_site_col(data)
}
\arguments{
\item{data}{Site metadata as obtained from \code{\link{dl_metadata}}}
}
\value{
A data frame with variables in one column and results of checks in
  the following columns
}
\description{
\code{qc_site_col} checks if all the site metadata columns were
created and their classes setted correctly
}
\details{
First qaulity check before continue to more complex checks. Here we are
simply testing if all the columns are there with the correct classes
(numerical for numerical values, character for text fields, etc.).
}

