% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_data_general_checks.R
\name{qc_solar_timestamp}
\alias{qc_solar_timestamp}
\title{Solar time conversion}
\usage{
qc_solar_timestamp(data, site_md, env_md, type = "apparent",
  parent_logger = "test")
}
\arguments{
\item{data}{Data frame containing the TIMESTAMP variable to convert to solar
time}

\item{site_md}{Data frame containing the latitude and longitude variables of
the site (\code{si_lat} and \code{si_long})}

\item{env_md}{Data frame containing the tz variable (\code{env_time_zone})}

\item{type}{Character indicating which kind of solar time is desired,
\code{mean} or \code{apparent}.}
}
\value{
A data frame exactly as \code{data}, but with the TIMESTAMP converted
  to the indicated solar time.
}
\description{
Calculate the Mean Solar Time and Apparent (Real) Solar Time from the
TIMESTAMP
}
\details{
In order to obtain the mean solar time and the equation of time for each day
included in the TIMESTAMP functions from \code{solaR} package.
}
\section{Apparent (Real) Solar Time}{

The Apparent Solar Time is calculated as:
\deqn{Apparent Solar Time = Mean Solar Time - Equation of Time}
The Equation of Time is calculated for each day, whereas the Mean Solar Time
is calculated for each step of the TIMESTAMP.
}
\seealso{
Other Quality Checks Functions: \code{\link{create_dic}},
  \code{\link{qc_coordinates}},
  \code{\link{qc_data_results_table}},
  \code{\link{qc_env_vars_presence}},
  \code{\link{qc_factor_values}},
  \code{\link{qc_fix_timestamp}},
  \code{\link{qc_get_sapw_md}},
  \code{\link{qc_get_timestep}},
  \code{\link{qc_get_timezone}},
  \code{\link{qc_is_timestamp}}, \code{\link{qc_md_cols}},
  \code{\link{qc_md_results_table}},
  \code{\link{qc_mind_the_gap_eff}},
  \code{\link{qc_mind_the_gap}},
  \code{\link{qc_pl_treatments}},
  \code{\link{qc_sapf_range_check}},
  \code{\link{qc_sapw_area_calculator}},
  \code{\link{qc_sapw_conversion}},
  \code{\link{qc_set_timezone}},
  \code{\link{qc_species_names_info}},
  \code{\link{qc_species_names}},
  \code{\link{qc_species_verification}},
  \code{\link{qc_start_process}},
  \code{\link{qc_time_interval}},
  \code{\link{qc_timestamp_concordance}},
  \code{\link{qc_timestamp_errors}}
}

