% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_species_functions.R
\name{qc_species_names_fix}
\alias{qc_species_names_fix}
\title{Verify provided species names (spelling and correctness)}
\usage{
qc_species_names_fix(data, parent_logger = "test")
}
\arguments{
\item{data}{Data frame as the obtained from \code{\link{qc_species_names_info}}}
}
\value{
A character vector with species fixed in spelling and correctness.
}
\description{
\code{\link[=qc_species_names]{qc_species_names()}} uses
(\href{https://matildabrown.github.io/rWCVP/index.html}{\code{rWCVP}}) package to
verify the species introduced by the contributors and fix spelling errors.
}
\details{
This function takes a vector of species names and check if they are right
spelled. Also, if a synonym is used, the function changes it automatically in
order to have the same name for the same species.
}
\section{GitHub package}{

\code{rWCVP} needs the companion \code{rWCVPdata} package, wich is a GitHub package.
It is not available
for install in the ususal way. In order to achieve that
\code{qc_species_verification} works as expected, manual installation of
this package must be done previously:
\code{devtools::install_github("matildabrown/rWCVPdata")}
}

