% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_md_general_checks.R
\name{qc_stand_dics}
\alias{qc_stand_dics}
\title{Dictionary creation for stand_md variables}
\usage{
qc_stand_dics(variable, parent_logger = "test")
}
\arguments{
\item{variable}{Variable name in which the dictionary is needed as character
vector (e.g. \code{'st_aspect'}).}
}
\value{
A character vector containing the valid values for the provided
variable
}
\description{
\code{qc_stand_dics} function creates a dictionary for the selected variable
containing the accepted values for that variable
}
\details{
In order to check if factor variables have a valid value or have been bad
formatted/introduced in the data template, first it is needed to have a list
of accepted values for each variable. This function creates that list to
use in the checks.
}
\section{Accepted variables}{

The factor variables in stand_md are \code{st_growth_condition},
\code{st_aspect}, \code{st_terrain} and \code{st_soil_texture}.
}

\seealso{
Other Dictionaries: 
\code{\link{qc_env_dics}()},
\code{\link{qc_plant_dics}()},
\code{\link{qc_site_dics}()},
\code{\link{qc_species_dics}()}
}
\concept{Dictionaries}
