% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_numerical_metadata_functions.R
\name{qc_suitable_range}
\alias{qc_suitable_range}
\title{Find if provided values of a numerical variable are in a suitable range}
\usage{
qc_suitable_range(data, variables, ranges)
}
\arguments{
\item{data}{Data frame containing the variables to check}

\item{variables}{Character vector with variables names}

\item{ranges}{List in the form of
\code{list("variable1" = c(min, max), "variable2" = c(min, max)...)}
indicating the suitable ranges for every variable contained in
\code{variable} (See details).}
}
\value{
A logical vector indicating if variables are in range
}
\description{
\code{qc_suitable_range} is used to determine if a variable has strange
values out of natural or biological expected range.
}
\details{
Checks are made in function of the variable and acceptable ranges provided.
\code{variables} and \code{ranges} must be of the same length, as the first
variable is compared with the first range provided, and so on. If both
objects are not of the same length, an informative error appears and function
stops.
}
\seealso{
Other Quality Check Functions: \code{\link{qc_check_coordinates}},
  \code{\link{qc_coord_sign_test}},
  \code{\link{qc_download_maps}},
  \code{\link{qc_fix_latlong_errors}}
}

