% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_unit_transformations.R
\name{qc_transf_list}
\alias{qc_transf_list}
\title{Transformations list}
\usage{
qc_transf_list(transf_info, parent_logger = "test")
}
\arguments{
\item{transf_info}{Data frame with info about the variables needed for
transformations as generated by \code{\link{qc_transformation_vars}}}
}
\value{
A data frame with the following columns:

\itemize{
\item{Transformation: Transformation/Conversion name}
\item{Available: Logical indicating if the transformation is possible}
}
}
\description{
Show all the transformations indicating which ones can be done with the data
provided.
}
\details{
The data frame returned by this function is intended to use it in the next
level, allowing automatized transformations depending on the available
variables.
}
\seealso{
Other Unit conversion: 
\code{\link{qc_cm_cm_h}()},
\code{\link{qc_rh}()},
\code{\link{qc_transformation_vars}()},
\code{\link{qc_units_process}()},
\code{\link{qc_vpd}()}
}
\concept{Unit conversion}
