% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_unit_transformations.R
\name{qc_vpd}
\alias{qc_vpd}
\title{Calculation of VPD from rh and ta}
\usage{
qc_vpd(data, parent_logger = "test")
}
\arguments{
\item{data}{Data frame containing the environmental data}
}
\value{
a Data frame of the environmental data with the new VPD variable
calculated
}
\description{
Calculate the VPD if needed from other environmental variables
}
\details{
The calculations for this function were obtained from the REddyProc R package
(https://cran.r-project.org/package=REddyProc) by the MPI-BGC in Jena
licensed under GPL > 2
}
\seealso{
Other Unit conversion: 
\code{\link{qc_cm_cm_h}()},
\code{\link{qc_rh}()},
\code{\link{qc_transf_list}()},
\code{\link{qc_transformation_vars}()},
\code{\link{qc_units_process}()}
}
\concept{Unit conversion}
