% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_functions.R
\name{rep_sfn_render}
\alias{rep_sfn_render}
\title{Render the report}
\usage{
rep_sfn_render(
  template,
  output_file = NULL,
  output_dir = NULL,
  parent_logger = "test",
  ...
)
}
\arguments{
\item{template}{Character vector indicating the name of the template. It must
be located in the Templates folder of the project}

\item{output_file}{Character vector indicating the name of the report file/s}

\item{output_dir}{Character vector indicating the output folder}

\item{...}{Parameters to pass on the yaml preamble of Rmd template}
}
\value{
Nothing, report is generated in the Report folder (default) or the
specified Report subfolder
}
\description{
Wrap around rmarkdown::render function to generate the automatic reports
}
\concept{Report Functions}
