% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_utils.R
\name{sfn_data2csv}
\alias{sfn_data2csv}
\title{sfn_data2csv function}
\usage{
sfn_data2csv(sfn_data, csv_folder)
}
\description{
This function is used in lvl3_process to write the csv files for each sfn_data objet
slots in the corresponding folder of the database tree
}
