% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{sfn_data_constructor}
\alias{sfn_data_constructor}
\title{Create a SfnData object from results of Quality Checks}
\usage{
sfn_data_constructor(sapf_data = NULL, env_data = NULL, site_md = NULL,
  stand_md = NULL, species_md = NULL, plant_md = NULL, env_md = NULL)
}
\arguments{
\item{sapf_data}{}

\item{env_data}{}

\item{site_md}{}

\item{stand_md}{}

\item{species_md}{}

\item{plant_md}{}

\item{env_md}{}
}
\value{
A SfnData object with all the data and metadata of the site
}
\description{
This function gather all the data and metadata generated in the QC process
and build an SfnData class object
}
\seealso{
Other Data Flow: \code{\link{df_accepted_to_lvl1}},
  \code{\link{df_copy_templates}},
  \code{\link{df_folder_structure}},
  \code{\link{df_get_data_folders}},
  \code{\link{df_get_status}},
  \code{\link{df_received_to_accepted}},
  \code{\link{df_report_folder_creation}},
  \code{\link{df_reset_data_status}},
  \code{\link{df_set_status}},
  \code{\link{df_start_status}}
}

