% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SfnData.R
\docType{data}
\name{sfn_get_generics}
\alias{get_env}
\alias{get_env_flags}
\alias{get_env_md}
\alias{get_env_timestamp}
\alias{get_plant_md}
\alias{get_sapf}
\alias{get_sapf_flags}
\alias{get_sapf_timestamp}
\alias{get_si_code}
\alias{get_site_md}
\alias{get_species_md}
\alias{get_stand_md}
\alias{sfn_get_generics}
\title{SfnData custom get generics}
\format{An object of class \code{character} of length 1.}
\usage{
get_sapf

get_env

get_sapf_flags

get_env_flags

get_sapf_timestamp

get_env_timestamp

get_site_md

get_stand_md

get_species_md

get_plant_md

get_env_md

get_si_code
}
\description{
Generics for getting the info in the slots of SfnData
}
\keyword{datasets}

