% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SfnData.R
\name{sfn_get_methods}
\alias{sfn_get_methods}
\title{SfnData get methods}
\arguments{
\item{object}{Object of class SfnData from which data is retrieved}

\item{...}{Further arguments to pass on to methods}
}
\description{
Methods to get the info from the SfnData class slots
}

