% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SfnData_methods.R
\name{sfn_replacement}
\alias{sfn_replacement}
\alias{get_sapf<-,SfnData-method}
\alias{get_env<-,SfnData-method}
\alias{get_sapf_flags<-,SfnData-method}
\alias{get_env_flags<-,SfnData-method}
\alias{get_timestamp<-,SfnData-method}
\alias{get_solar_timestamp<-,SfnData-method}
\alias{get_si_code<-,SfnData-method}
\alias{get_site_md<-,SfnData-method}
\alias{get_stand_md<-,SfnData-method}
\alias{get_species_md<-,SfnData-method}
\alias{get_plant_md<-,SfnData-method}
\alias{get_env_md<-,SfnData-method}
\title{Replacement methods}
\usage{
\S4method{get_sapf}{SfnData}(object) <- value

\S4method{get_env}{SfnData}(object) <- value

\S4method{get_sapf_flags}{SfnData}(object) <- value

\S4method{get_env_flags}{SfnData}(object) <- value

\S4method{get_timestamp}{SfnData}(object) <- value

\S4method{get_solar_timestamp}{SfnData}(object) <- value

\S4method{get_si_code}{SfnData}(object) <- value

\S4method{get_site_md}{SfnData}(object) <- value

\S4method{get_stand_md}{SfnData}(object) <- value

\S4method{get_species_md}{SfnData}(object) <- value

\S4method{get_plant_md}{SfnData}(object) <- value

\S4method{get_env_md}{SfnData}(object) <- value
}
\value{
The same SfnData object with the corresponding slot changed to the
value provided. An error if the value provided generates an invalid
SfnData object.
}
\description{
Methods for replacing the slots with new data or metadata
}
\details{
The replacement object must be a valid object for that slot, i.e. for sapflow
data slot a data frame with the same dimensions and without TIMESTAMP variable
is needed. A validity check is done before returning the replaced SfnData
object and an error is returned if this check fails.
}
