% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SfnData_class.R
\docType{class}
\name{SfnData-class}
\alias{SfnData}
\alias{SfnData-class}
\title{An S4 class for sapfluxnet data}
\description{
An S4 class for sapfluxnet data
}
\section{Slots}{

\describe{
\item{\code{sapf}}{A data frame with the sapf data}

\item{\code{env}}{A data frame with the env data}

\item{\code{sapf_flags}}{A data frame with the same dimensions of \code{sapf} with the
flag info for each tree/TIMESTAMP combination}

\item{\code{env_flags}}{A data frame with the same dimensions of \code{env} with the
flag info for each tree/TIMESTAMP combination}

\item{\code{si_code}}{A character indicating the site code}

\item{\code{site_md}}{A data frame containing the site metadata}

\item{\code{stand_md}}{A data frame containing the stand metadata}

\item{\code{species_md}}{A data frame containing the species metadata}

\item{\code{plant_md}}{A data frame containing the plant metadata}

\item{\code{env_md}}{A data frame containing the env metadata}
}}

