% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SfnData.R
\docType{methods}
\name{show,SfnData-method}
\alias{show,SfnData-method}
\title{Show method for SfnData}
\usage{
\S4method{show}{SfnData}(object)
}
\arguments{
\item{object}{SfnData object to show}
}
\description{
Show method for SfnData
}

