% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_utils.R
\name{sm_solarTIMESTAMP_adder}
\alias{sm_solarTIMESTAMP_adder}
\title{solarTIMESTAMP adder}
\usage{
sm_solarTIMESTAMP_adder(si_code, parent_logger = "test")
}
\arguments{
\item{si_code}{Character indicating the site code}
}
\value{
Invisible TRUE if solarTIMESTAMP was succesfully added to all levels
in si_code site
}
\description{
Add solarTIMESTAMP slot to SfnData objects
}
\details{
solarTIMESTAMP was a latter addition to the SfnData class. Sites ran with the
pre-solarTIMESTAMP version lacks this slot, which will cause problems in the
data flow. This function looks for level 1 SfnData object and update it.
}
\seealso{
Other Server Management: 
\code{\link{sm_status_updater}()}
}
\concept{Server Management}
