% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SfnData_methods.R
\docType{methods}
\name{[,SfnData-method}
\alias{[,SfnData-method}
\title{Sub-setting operation}
\usage{
\S4method{[}{SfnData}(x, i, j, k, drop = "missing")
}
\arguments{
\item{i}{sapflow data row index}

\item{j}{sapflow data column index}

\item{k}{env data row index}

\item{l}{env data column index}

\item{object}{SfnData object}
}
\description{
Sub-setting operation
}

