% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SfnData_methods.R
\name{[,SfnData,numeric,ANY,missing-method}
\alias{[,SfnData,numeric,ANY,missing-method}
\title{Sub-setting operation}
\usage{
\S4method{[}{SfnData,numeric,ANY,missing}(x, i, j, k, drop)
}
\arguments{
\item{i}{data row index}

\item{j}{sapflow data column index}

\item{k}{env data column index}

\item{object}{SfnData object}
}
\description{
Sub-setting operation
}
