% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_functions.R
\name{theme_sfn}
\alias{theme_sfn}
\title{ggplot2 theme for SAPFLUXNET plots}
\usage{
theme_sfn(base_size = 10, base_family = "Lato")
}
\description{
Custom ggplot2 theme for uniformization of plot visuals
}
