% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_functions.R
\name{vis_biome}
\alias{vis_biome}
\title{Plotting a diagram of biomes}
\usage{
vis_biome(merge_deserts = FALSE, parent_logger = "test")
}
\arguments{
\item{merge_deserts}{Logical indicating if desert biomes should be merged
in a single biome. By default, deserts are not merged.}
}
\value{
a ggplot object showing the biomes.
}
\description{
This function produces a ggplot object showing the biomes as colored areas
according to mean annual temperature (MAT) and mean annual precipitation (MAP)
using a SpatialPolygonsDataFrame object obtained with
\code{\link{qc_get_biomes_spdf}}
}
\seealso{
Other Visualization Functions: \code{\link{vis_gap_lines}},
  \code{\link{vis_location_biome}},
  \code{\link{vis_plot_the_gap}}
}

