% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_functions.R
\name{vis_environmental_responses}
\alias{vis_environmental_responses}
\title{Environmental responses plot}
\usage{
vis_environmental_responses(SfnData, env_var = "vpd", solar = FALSE,
  parent_logger = "test")
}
\arguments{
\item{SfnData}{SfnData object}

\item{env_var}{Character indicating the nameof the environmental variable to
plot}

\item{solar}{Use solarTIMESTAMP?}
}
\value{
a \code{ggplot} object with the desired plot
}
\description{
Plot the desired environmental funcion \emph{vs.} spaflow values
}
\seealso{
Other Visualization Functions: \code{\link{vis_biome}},
  \code{\link{vis_gap_lines}},
  \code{\link{vis_location_biome}},
  \code{\link{vis_plot_the_gap}}
}
