% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_functions.R
\name{vis_gap_lines}
\alias{vis_gap_lines}
\title{TIMESTAMP with gaps visualization}
\usage{
vis_gap_lines(sapf_data = NULL, env_data = NULL, parent_logger = "test")
}
\arguments{
\item{sapf_data}{Data frame with the sapflow data}

\item{env_data}{Data frame with the environmental data}
}
\value{
A ggplot object with the basic lines plot, no themes added.
}
\description{
Concordance lines plot with gaps
}
\seealso{
Other Visualization Functions: \code{\link{vis_biome}},
  \code{\link{vis_location_biome}},
  \code{\link{vis_plot_the_gap}}
}

