% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_functions.R
\name{vis_location_biome}
\alias{vis_location_biome}
\title{Plotting a diagram of biomes with sites as dots}
\usage{
vis_location_biome(data, merge_deserts = FALSE, parent_logger = "test")
}
\arguments{
\item{data}{Data frame of site metadata, including mean annual temperature
(si_mat) and mean annual precipitation (si_map) columns, or at least
latitude (si_lat) and longitude (si_long) columns that will be used to obtain
climatic data with \code{\link{qc_get_biome}}.}

\item{merge_deserts}{Logical indicating if desert biomes should be merged
in a single biome. By default, deserts are not merged.}
}
\value{
a ggplot object showing the biomes.
}
\description{
This function produces a ggplot object showing the biomes as colored areas
according to mean annual temperature (MAT) and mean annual precipitation (MAP),
using the function \code{\link{vis_biome}}, and adds the sites on it according
to their values of MAT and MAP.
}
\seealso{
Other Visualization Functions: \code{\link{vis_biome}},
  \code{\link{vis_gap_lines}},
  \code{\link{vis_plot_the_gap}}
}

