% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_functions.R
\name{vis_plot_the_gap}
\alias{vis_plot_the_gap}
\title{Plotting an histogram for gaps intervals/gaps coverage}
\usage{
vis_plot_the_gap(gaps_info, type = "gap_interval", binwidth = NULL,
  parent_logger = "test")
}
\arguments{
\item{gaps_info}{Data frame as obtained from \code{\link{qc_mind_the_gap}}}

\item{type}{Character indicating what to represent, \code{gap_interval} or
\code{gap_coverage}}

\item{binwidth}{Bin width as stated in geom_histogram, default to NULL to
use the geom_histrogram default. Change it if more or less resolution is
needed. Only works for \code{type = 'gap_interval'}.}
}
\value{
a ggplot object with the basic histogram, no themes added.
}
\description{
Wrapper for ggplot to plot an histogram of gaps info
}
\details{
This function is a simple wrapper for ggplot + geom_histogram. It produces
a ggplot object that can be modified adding layers, like any other ggplot
object.
}
\seealso{
Other Visualization Functions: \code{\link{vis_biome}},
  \code{\link{vis_gap_lines}},
  \code{\link{vis_location_biome}}
}
