% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_utils.R
\name{write_sfn_data}
\alias{write_sfn_data}
\title{Function to save the sfn_data created}
\usage{
write_sfn_data(sfn_data, folder, parent_logger = "test")
}
\description{
This function saves the RData files with the sfn_data objects
}
