% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_md_general_checks.R
\name{create_dic}
\alias{create_dic}
\title{Variable dictionaries for metadata}
\usage{
create_dic(dic_name)
}
\arguments{
\item{dic_name}{Name of the metadata sheet of which dictionary is needed.
It must be one of \code{site_md}, \code{stand_md}, \code{species_md},
\code{plant_md} or \code{environmental_md}.}
}
\value{
A named list, variable names being the index and class the value
}
\description{
\code{create_dic} creates a dictionary containing metadata variable names
and expected class. It is intended as an internal function.
}
\details{
After loading metadata sheets, introduced variables and their classes must be
checked in order to ensure data correctness. For that, we need dictionaries
containing all the expected variables and their corresponding classes to
compare. This function works inside of \code{\link{qc_col_class}}, so there
is no need to call it directly
}
\seealso{
Other Quality Checks Functions: \code{\link{qc_md_cols}},
  \code{\link{qc_species_verification}}
}

