% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{df_accepted_to_lvl1}
\alias{df_accepted_to_lvl1}
\title{Save the fixed datasets (metadata and data) in the LVL1 folder}
\usage{
df_accepted_to_lvl1(si_code, sapf_data_plant = NULL,
  sapf_data_sapwood = NULL, sapf_data_leaf = NULL, env_data = NULL,
  site_md = NULL, stand_md = NULL, plant_md = NULL, species_md = NULL,
  env_md = NULL, rdata = FALSE, parent_logger = "test")
}
\arguments{
\item{si_code}{Character with the site code}

\item{env_data}{Data frame with the fixed environmental data}

\item{site_md}{Data frame with the fixed site metadata}

\item{stand_md}{Data frame with the fixed stand metadata}

\item{plant_md}{Data frame with the fixed plant metadata}

\item{species_md}{Data frame with the fixed species metadata}

\item{env_md}{Data frame with the fixed environmental metadata}

\item{rdata}{Logical indicating if an RData file must be created with the
objects generated in the QC. Default to FALSE.}

\item{sapf_data}{Data frame with the fixed sapflow data}
}
\value{
Nothing
}
\description{
This function performs three actions: 1) write csv files with the fixed data
in the LVL1 folder. 2) Update the status files to indicate that data is in
LVL1 and the date of the move. 3) Save the objects generated in the QC to
make easy the creation of dashboards without the need of run all QC tests
again.
}
\seealso{
Other Data Flow: \code{\link{df_copy_templates}},
  \code{\link{df_folder_structure}},
  \code{\link{df_get_data_folders}},
  \code{\link{df_get_status}},
  \code{\link{df_received_to_accepted}},
  \code{\link{df_report_folder_creation}},
  \code{\link{df_reset_data_status}},
  \code{\link{df_set_status}},
  \code{\link{df_start_status}},
  \code{\link{sfn_data_constructor}}
}

