% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{df_get_status}
\alias{df_get_status}
\title{Get the status file info}
\usage{
df_get_status(si_code, parent_logger = "test")
}
\arguments{
\item{si_code}{Character vector indicating the site code}
}
\value{
a list object with the info contained in the status file
}
\description{
Retrieve the status file info as a list, with the yaml package
}
\details{
\code{yaml} file is parsed by yaml.load_file and a list object is returned.
}
\seealso{
Other Data Flow: \code{\link{df_accepted_to_lvl1}},
  \code{\link{df_copy_templates}},
  \code{\link{df_folder_structure}},
  \code{\link{df_get_data_folders}},
  \code{\link{df_lvl1_to_lvl2}},
  \code{\link{df_lvl2_folder_structure}},
  \code{\link{df_read_SfnData}},
  \code{\link{df_received_to_accepted}},
  \code{\link{df_rem_to_units}},
  \code{\link{df_report_folder_creation}},
  \code{\link{df_reset_data_status}},
  \code{\link{df_set_status}},
  \code{\link{df_start_status}},
  \code{\link{df_warn_to_rem}},
  \code{\link{df_whos_ready_to}},
  \code{\link{df_write_SfnData}},
  \code{\link{sfn_data_constructor}}
}
