% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{df_received_to_accepted}
\alias{df_received_to_accepted}
\title{Move from received to accepted data folders}
\usage{
df_received_to_accepted(remove = FALSE)
}
\arguments{
\item{remove}{Logical indicating if files in received folder are dropped after
correct after the file transfer. Default to FALSE.}
}
\value{

}
\description{
Check if data file/s already exists and create the folder structure needed to
store them.
}
\details{
The first step after receiving a data set is to move it from received folder
to accepted folder (\code{Data/site_code/site_code_accepted/}).
}
\seealso{
Other Data Flow: \code{\link{df_folder_structure}}
}

