% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{df_received_to_accepted}
\alias{df_received_to_accepted}
\title{Move from received to accepted data folders}
\usage{
df_received_to_accepted(remove = FALSE, parent_logger = "test")
}
\arguments{
\item{remove}{Logical indicating if files in received folder are dropped after
correct after the file transfer. Default to FALSE.}
}
\value{

}
\description{
Check if data file/s already exists and create the folder structure needed to
store them.
}
\details{
The first step after receiving a data set is to move it from received folder
to accepted folder (\code{Data/site_code/site_code_accepted/}).
}
\seealso{
Other Data Flow: \code{\link{df_accepted_to_lvl1}},
  \code{\link{df_copy_templates}},
  \code{\link{df_folder_structure}},
  \code{\link{df_get_data_folders}},
  \code{\link{df_get_status}},
  \code{\link{df_lvl1_to_lvl2}},
  \code{\link{df_lvl2_folder_structure}},
  \code{\link{df_read_SfnData}},
  \code{\link{df_report_folder_creation}},
  \code{\link{df_reset_data_status}},
  \code{\link{df_set_status}},
  \code{\link{df_start_status}},
  \code{\link{df_warn_to_rem}},
  \code{\link{df_whos_ready_to}},
  \code{\link{df_write_SfnData}},
  \code{\link{sfn_data_constructor}}
}
