% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{df_report_folder_creation}
\alias{df_report_folder_creation}
\title{Reports folders creation}
\usage{
df_report_folder_creation(si_code, parent_logger = "test")
}
\arguments{
\item{si_code}{Character indicating the site code to create the reports folder}
}
\value{
Invisible TRUE if the folder is created correctly, invisible FALSE if
  folder is not created.
}
\description{
Function to create the report subfolder for the site to analyze
}
\seealso{
Other Data Flow: \code{\link{df_accepted_to_lvl1}},
  \code{\link{df_copy_templates}},
  \code{\link{df_folder_structure}},
  \code{\link{df_get_data_folders}},
  \code{\link{df_get_status}},
  \code{\link{df_lvl1_to_lvl2}},
  \code{\link{df_lvl2_folder_structure}},
  \code{\link{df_read_SfnData}},
  \code{\link{df_received_to_accepted}},
  \code{\link{df_reset_data_status}},
  \code{\link{df_set_status}},
  \code{\link{df_start_status}},
  \code{\link{df_who_ready_to_lvl2}},
  \code{\link{df_write_SfnData}},
  \code{\link{sfn_data_constructor}}
}
