% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{df_who_ready_to_lvl2}
\alias{df_who_ready_to_lvl2}
\title{Who is ready for level 2?}
\usage{
df_who_ready_to_lvl2(parent_logger = "test")
}
\value{
A list with length equal to the number of sites containing the
  TO_LVL2 flag of the status files.
}
\description{
Check the site status files to list who is ready to move to level 2
}
\seealso{
Other Data Flow: \code{\link{df_accepted_to_lvl1}},
  \code{\link{df_copy_templates}},
  \code{\link{df_folder_structure}},
  \code{\link{df_get_data_folders}},
  \code{\link{df_get_status}},
  \code{\link{df_lvl2_folder_structure}},
  \code{\link{df_received_to_accepted}},
  \code{\link{df_report_folder_creation}},
  \code{\link{df_reset_data_status}},
  \code{\link{df_set_status}},
  \code{\link{df_start_status}},
  \code{\link{sfn_data_constructor}}
}

