% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{df_whos_ready_to}
\alias{df_whos_ready_to}
\title{Who is ready the desired level?}
\usage{
df_whos_ready_to(level = c("lvl2", "rem", "units"), filter = c("all",
  "ready", "freeze", "done"), parent_logger = "test")
}
\arguments{
\item{level}{string indicating the level to check. Accepted values are
"lvl2", "rem" and "units".}

\item{filter}{character vector indicating by which status results must
been filtered. Accepted values are "all" (default), "ready", "freeze" or "done"
(see details)}
}
\value{
A list with length equal to the number of sites containing the
  \code{TO_[level]} flag of the status files.
}
\description{
Check the site status files to list who is ready to move to target
}
\details{
This function check the status files of all sites and list the status
for the desired level.
\code{filter} parameter can be used to filter the results by any status:

\itemize{
    \item{\code{all} retrieves all the statuses}
    \item{\code{ready} retrieves only those sites marked to pass to level 2}
    \item{\code{freeze} retrieves only those sites freezed in level 1 yet}
    \item{\code{done} retrieves only those sites already passed to level 2}
  }
}
\seealso{
Other Data Flow: \code{\link{df_accepted_to_lvl1}},
  \code{\link{df_copy_templates}},
  \code{\link{df_folder_structure}},
  \code{\link{df_get_data_folders}},
  \code{\link{df_get_status}},
  \code{\link{df_lvl1_to_lvl2}},
  \code{\link{df_lvl2_folder_structure}},
  \code{\link{df_read_SfnData}},
  \code{\link{df_received_to_accepted}},
  \code{\link{df_rem_to_units}},
  \code{\link{df_report_folder_creation}},
  \code{\link{df_reset_data_status}},
  \code{\link{df_set_status}},
  \code{\link{df_start_status}},
  \code{\link{df_warn_to_rem}},
  \code{\link{df_write_SfnData}},
  \code{\link{sfn_data_constructor}}
}
