% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_data_loading_functions.R
\name{dl_data}
\alias{dl_data}
\title{Loading data from xls/xlsx}
\usage{
dl_data(file_name, sheet_name, long = FALSE)
}
\arguments{
\item{file_name}{Character vector indicating the name of the xls/xlsx file
containing the data.}

\item{sheet_name}{Character vector indicating the name of the sheet to be
loaded. It must be one of \code{sapflow_hd} or \code{environmental_hd}.}

\item{long}{Logical indicating if returned data must be in \code{long} format}
}
\value{
\code{dl_data} returns sapflow or environmental data in wide format,
  ready to pipe it in the quality checks for raw data. If \code{long = TRUE}
  data is returned in long format, ready for plotting.
}
\description{
This function make use of dplyr, tidyr and readxl packages in order to
retrieve and format the data.
}
\details{
In case of data to be in the xls/xlsx file, instead of separate csv files,
this function is used to retrieve the raw data (sapflow and environmental).
}
\seealso{
Other Data Loading Functions: \code{\link{dl_metadata}},
  \code{\link{remove_dupcols}}
}

