% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_data_loading_functions.R
\name{dl_dec_char_detect}
\alias{dl_dec_char_detect}
\title{Decimal character detection}
\usage{
dl_dec_char_detect(file, n = 1000, parent_logger = "test")
}
\arguments{
\item{file}{Character vector indicating the name (and route) of the file in
which the decimal character must be guessed.}

\item{n}{Numeric indicating the sample size of rows to use in the decimal
character guessing.}
}
\value{
A character vector containing the decimal character, to use in the
  fread order to load the data.
}
\description{
Decimal character detection to csv files loading with data.table::fread
}
\details{
With data coming from any place in the world, data loading functions have to
cope with different decimal characters, separator characters...
\code{\link[data.table]{fread}} can autodetect separator character at loading
data, but decimal character must be specified in order to load correct data.
This function aims to provide an easy way to autodetect decimal character to
automatize data loading proccess.
}
\section{Algorithm}{

The process to determine the decimal character is as follows:
\enumerate{
  \item 1000 rows and 4 variables are sampled from the data.
  \item For each variable, number of dots (\code{.}) and commas (\code{,})
  are calculated.
  \item If the number of dots is larger than the commas, dot is selected as
  decimal character. If the number of commas is larger than the dots, comma
  is selected as decimal character.
}
}

\seealso{
Other Data Loading Functions: 
\code{\link{dl_data}()},
\code{\link{dl_data_col_classes}()},
\code{\link{dl_get_si_code}()},
\code{\link{dl_metadata}()},
\code{\link{dl_na_char_generator}()},
\code{\link{qc_as_timestamp}()},
\code{\link{remove_dupcols}()}
}
\concept{Data Loading Functions}
