% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_data_loading_functions.R
\name{dl_na_char_generator}
\alias{dl_na_char_generator}
\title{Create a characters vector for NA values}
\usage{
dl_na_char_generator(parent_logger = "test")
}
\value{
A character vector with the text strings to be converted to NAs, in
  order to use it in data.table::fread function.
}
\description{
Generation of a character vector with the possible na characters found in
csv files in order to load data with data.table::fread
}
\details{
Sometimes, csv files exported from excel files have some characters in the
cells, causing data columns to be interpreted as characters or factors
instead of numeric. This function compile an exhaustive list of those
characters in order to transform them in NAs at the moment of reading the
data.
}
\seealso{
Other Data Loading Functions: 
\code{\link{dl_data}()},
\code{\link{dl_data_col_classes}()},
\code{\link{dl_dec_char_detect}()},
\code{\link{dl_get_si_code}()},
\code{\link{dl_metadata}()},
\code{\link{qc_as_timestamp}()},
\code{\link{remove_dupcols}()}
}
\concept{Data Loading Functions}
