% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_coordinates_functions.R
\name{download_maps}
\alias{download_maps}
\title{Download maps for countries included in the database}
\usage{
download_maps(data, folder = getwd())
}
\arguments{
\item{data}{Data frame where the countries ISO codes are. Must contain a
variable called \code{country} where the ISO code resides.}

\item{folder}{Folder route where the maps are stored or where they will be
stored, by default the working directory. It must be a character object and
it must end \bold{without} \code{/}.}
}
\value{
Maps are downloaded (if needed) and a summary is returned indicating
  number of maps downloaded and number of maps present in the map folder.
}
\description{
\code{download_maps} fetch maps from \url{http://www.gadm.org/}.
}
\details{
This function fetch maps from \url{http://www.gadm.org/} if the map is not
already present in the maps folder (by default, the working directory).
}
\seealso{
Other Quality Check Functions: \code{\link{check_coordinates}},
  \code{\link{coord_sign_test}},
  \code{\link{fix_latlong_errors}},
  \code{\link{qc_suitable_range}}
}

