% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control_functions.R
\name{fix_latlong_errors}
\alias{fix_latlong_errors}
\title{Fixing sign errors in coordinates}
\usage{
fix_latlong_errors(data, maps_folder, sign_errors = TRUE)
}
\arguments{
\item{data}{Data frame with data coming from \code{\link{check_coordinates}}
(with latitude, longitude, country and is_inside_country variables).}

\item{maps_folder}{Folder route where the maps are stored, by default the
working directory. It must be a character object and it must end
\bold{without} \code{/}.}

\item{sign_errors}{Logical indicating if sign errors must be checked and
fixed. If TRUE (default), \code{\link{coord_sign_test}} is internally
called.}
}
\value{
Same data frame provided, with coordinates tested and fixed
}
\description{
\code{fix_latlong_errors} makes possible to fix known errors in latitude and
longitude coordinates, as exchanged signs.
}
\details{
This function calls to other internal functions in order to fix different
kinds of coordinates errors. At the moment, only exchanged signs in
coordinates errors are allowed to be fixed.
If \code{sign_errors = TRUE} is specified, \code{\link{coord_sign_test}} is
called to establish possible sign error and, if any, they are fixed.
}
\seealso{
Other Quality.Check.Functions: \code{\link{check_coordinates}},
  \code{\link{coord_sign_test}},
  \code{\link{download_maps}}
}

