% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_outliers_app.R
\name{out_app}
\alias{out_app}
\title{outliers_app function}
\usage{
out_app(parent_logger = "test")
}
\value{
An interactive app to manage outliers
}
\description{
Interactive app to flag outliers, out_of_range data and visually detected
data flaws.
}
\details{
This function launches a shiny app to visually manage flagging outliers,
out_of_range values and visual data flaws. Data values selected after
visual exploration of the data presented in the app are coded in a script
saved in the \code{Data/si_code/Lvl_2/visual_check} folder. This script
performs the data flagging needed and serves as a reproducible tool in case
of reconstruction of the complete process undergo by the site data.
}

