% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_data_general_checks.R
\name{qc_as_timestamp}
\alias{qc_as_timestamp}
\title{Fixing known errors in TIMESTAMP format}
\usage{
qc_as_timestamp(data, parent_logger = "test")
}
\arguments{
\item{data}{Data frame containing TIMESTAMP variable. Also it can be a vector
with TIMESTAMP values}
}
\value{
An object of the same type of input (data frame or vector) with the
  fixed values of TIMESTAMP. If TIMESTAMP is already in format, a message
  appears and none fix is made, returning data as entered.
  If TIMESTAMP can not be fixed, an error is raised.
}
\description{
Converting known bad TIMESTAMP formats to POSIXt
}
\details{
When loading data from csv files, depending on the office version and
workflow to introduce TIMESTAMP and data, TIMESTAMP can result in the wrong
format (lost of seconds, for example). This function checks for known
formatting errors and try to fix them.
}
\seealso{
Other Data Loading Functions: \code{\link{dl_data}},
  \code{\link{dl_dec_char_detect}},
  \code{\link{dl_metadata}},
  \code{\link{dl_na_char_generator}},
  \code{\link{remove_dupcols}}
}

