% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_coordinates_functions.R
\name{qc_check_coordinates}
\alias{qc_check_coordinates}
\title{Site coordinates checking}
\usage{
qc_check_coordinates(
  data,
  maps_folder = getwd(),
  plot = FALSE,
  text_report = TRUE,
  parent_logger = "test"
)
}
\arguments{
\item{data}{Data frame with data. At least, longitude, latitude, country and
site name variables must be present in the data object.}

\item{maps_folder}{Folder route where the maps are stored, by default the
working directory. It must be a character object and it must end
\bold{without} \code{/}.}

\item{plot}{Logical indicating if plots for coordinate are created and saved
in the working directory. By default, plot are not saved.}

\item{text_report}{Logical indicating if a text report is showed in the
console after checking coordinates. By default, a report is showed in the
console.}
}
\value{
The data frame used as input with a new variable, is_inside_country,
a logical variable indicating if the site has wrong coordinates, and another
new variable is_on_land, a logical variable with TRUE if the coordinates are
on land and false if they are over water bodies.
}
\description{
\code{qc_check_coordinates} verifies if provided coordinates are within
country declared in metadata form.
}
\details{
This function internally download the maps with
\code{\link{qc_download_maps}} and it checks if the provided coordinates are
within the country limits. It only creates a data frame containing country,
site and a logical variable indicating if coordinates are correct.
}
\seealso{
Other Quality Check Functions: 
\code{\link{qc_coord_sign_test}()},
\code{\link{qc_download_maps}()},
\code{\link{qc_fix_latlong_errors}()}
}
\concept{Quality Check Functions}
