% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_coordinates_functions.R
\name{qc_coordinates}
\alias{qc_coordinates}
\title{Commodity function to check site coordinates}
\usage{
qc_coordinates(data, maps_folder = getwd(), plot = FALSE,
  text_report = TRUE, sign_errors = TRUE, special_countries = TRUE,
  parent_logger = "test")
}
\arguments{
\item{data}{Data frame with data coming from \code{\link{qc_check_coordinates}}
(with latitude, longitude, country and is_inside_country variables).}

\item{maps_folder}{Folder route where the maps are stored, by default the
working directory. It must be a character object and it must end
\bold{without} \code{/}.}

\item{plot}{Logical indicating if plots for coordinate are created and saved
in the working directory. By default, plot are not saved.}

\item{text_report}{Logical indicating if a text report is showed in the
console after checking coordinates. By default, a report is showed in the
console.}

\item{sign_errors}{Logical indicating if sign errors must be checked and
fixed. If TRUE (default), \code{\link{qc_coord_sign_test}} is internally
called.}

\item{special_countries}{Logical indicating if the special approach to
countries having positive and negative coordinates must be used. See
\code{\link{qc_fix_latlong_errors}} for details.}
}
\description{
Wrap around \code{\link{qc_check_coordinates}} and \code{\link{qc_fix_latlong_errors}}
to process coordinates data in one step
}
\seealso{
Other Quality Checks Functions: \code{\link{create_dic}},
  \code{\link{qc_data_results_table}},
  \code{\link{qc_env_vars_presence}},
  \code{\link{qc_ext_radiation}},
  \code{\link{qc_factor_values}},
  \code{\link{qc_fix_timestamp}},
  \code{\link{qc_get_biomes_spdf}},
  \code{\link{qc_get_biome}}, \code{\link{qc_get_sapw_md}},
  \code{\link{qc_get_timestep}},
  \code{\link{qc_get_timezone}},
  \code{\link{qc_is_timestamp}}, \code{\link{qc_md_cols}},
  \code{\link{qc_md_results_table}},
  \code{\link{qc_mind_the_gap_eff}},
  \code{\link{qc_mind_the_gap}},
  \code{\link{qc_pl_treatments}},
  \code{\link{qc_rad_conversion}},
  \code{\link{qc_sapf_range_check}},
  \code{\link{qc_sapw_area_calculator}},
  \code{\link{qc_sapw_conversion}},
  \code{\link{qc_set_timezone}},
  \code{\link{qc_soil_texture}},
  \code{\link{qc_species_names_info}},
  \code{\link{qc_species_names}},
  \code{\link{qc_species_verification}},
  \code{\link{qc_start_process}},
  \code{\link{qc_time_interval}},
  \code{\link{qc_timestamp_concordance}},
  \code{\link{qc_timestamp_errors}},
  \code{\link{qc_timestamp_nas}}
}
