% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_main.R
\name{qc_data_results_table}
\alias{qc_data_results_table}
\title{Main function to resume Data QC in one data frame}
\usage{
qc_data_results_table(sapf_data_fixed, env_data_fixed, timestamp_errors_sapf,
  timestamp_errors_env, sapw_md, timestamp_concordance, sapf_gaps_info,
  env_gaps_info, sapf_timestamp_nas, env_timestamp_nas, swc_check,
  transformations_table, parent_logger = "test")
}
\arguments{
\item{sapf_data_fixed}{}

\item{env_data_fixed}{}

\item{timestamp_errors_sapf}{}

\item{timestamp_errors_env}{}

\item{sapw_md}{}

\item{timestamp_concordance}{}

\item{sapf_gaps_info}{}

\item{env_gaps_info}{}

\item{sapf_timestamp_nas}{}

\item{env_timestamp_nas}{}

\item{swc_check}{}

\item{transformations_table}{}
}
\description{
Data QC codified results in one data frame
}
\seealso{
Other Quality Checks Functions: \code{\link{create_dic}},
  \code{\link{qc_coordinates}},
  \code{\link{qc_env_ranges}},
  \code{\link{qc_env_vars_presence}},
  \code{\link{qc_ext_radiation}},
  \code{\link{qc_factor_values}},
  \code{\link{qc_fix_timestamp}},
  \code{\link{qc_get_biomes_spdf}},
  \code{\link{qc_get_biome}}, \code{\link{qc_get_sapw_md}},
  \code{\link{qc_get_timestep}},
  \code{\link{qc_get_timezone}},
  \code{\link{qc_is_timestamp}}, \code{\link{qc_md_cols}},
  \code{\link{qc_md_results_table}},
  \code{\link{qc_mind_the_gap_eff}},
  \code{\link{qc_mind_the_gap}},
  \code{\link{qc_out_hampel_filter}},
  \code{\link{qc_out_of_range}},
  \code{\link{qc_out_remove}},
  \code{\link{qc_outliers_process}},
  \code{\link{qc_outliers_subs}},
  \code{\link{qc_pl_treatments}},
  \code{\link{qc_rad_conversion}},
  \code{\link{qc_range_dic}}, \code{\link{qc_sapf_ranges}},
  \code{\link{qc_sapw_area_calculator}},
  \code{\link{qc_sapw_conversion}},
  \code{\link{qc_set_timezone}},
  \code{\link{qc_soil_texture}},
  \code{\link{qc_species_names_info}},
  \code{\link{qc_species_names}},
  \code{\link{qc_species_verification}},
  \code{\link{qc_start_process}},
  \code{\link{qc_swc_check}}, \code{\link{qc_swc_fix}},
  \code{\link{qc_time_interval}},
  \code{\link{qc_timestamp_concordance}},
  \code{\link{qc_timestamp_errors}},
  \code{\link{qc_timestamp_nas}}
}
