% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_sitemd_checks.R
\name{qc_email_check}
\alias{qc_email_check}
\title{Check if provided email directions are correct}
\usage{
qc_email_check(data, parent_logger = "test")
}
\arguments{
\item{data}{Data frame containing the site metadata where email variables are
located.}
}
\value{
A data frame with the directions and the result of the check
}
\description{
\code{qc_email_check} function checks for correctness in the email fields
}
\details{
Provided email direction/s (contact and additional contributor, if any) are
checked to ensure that are valid email directions. This function uses
stringr package utilities for character string tinkering and regex to match
the characteristics of email directions
}
\section{Regex}{

  Email directions can be divided in three components, \code{username},
  \code{@} and \code{domain}. \code{username} component can contain
  any letter, number, underscores, dots and hyphens. \code{domain} component
  can be also any letter, number, underscores dots and hyphens, followed by
  a dot and from two to six letters and/or dots. So, the regex expression
  to match a email direction is as follows:
  \code{"^([a-z0-9_\\.-]+)@([\\da-z\\.-]+)\\.([a-z\\.]{2,6})$"}
}

