% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_md_general_checks.R
\name{qc_env_dics}
\alias{qc_env_dics}
\title{Dictionary creation for environmental_md variables}
\usage{
qc_env_dics(variable, parent_logger = "test")
}
\arguments{
\item{variable}{Variable name in which the dictionary is needed as character
vector (e.g. \code{'env_vpd'}).}
}
\value{
A character vector containing the valid values for the provided
  variable
}
\description{
\code{qc_env_dics} function creates a dictionary for the selected variable
containing the accepted values for that variable
}
\details{
In order to check if factor variables have a valid value or have been bad
formatted/introduced in the data template, first it is needed to have a list
of accepted values for each variable. This function creates that list to
use in the checks.
}
\section{Accepted variables}{

The factor variables in environmental_md are \code{env_time_zone},
\code{env_ta}, \code{env_rh}, \code{env_vpd}, \code{env_sw_in},
\code{env_ppfd_in}, \code{env_netrad}, \code{env_ws}, \code{env_precip},
\code{env_plant_watpot} and \code{env_leafarea_seasonal}.
}

\seealso{
Other Dictionaries: \code{\link{qc_plant_dics}},
  \code{\link{qc_site_dics}},
  \code{\link{qc_species_dics}},
  \code{\link{qc_stand_dics}}
}
