% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_md_general_checks.R
\name{qc_env_vars_presence}
\alias{qc_env_vars_presence}
\title{Check for environmental variables presence}
\usage{
qc_env_vars_presence(data, metadata)
}
\arguments{
\item{data}{Data frame containing the environmental data}

\item{metadata}{Data frame containing the environmental metadata}
}
\value{
A data frame with four columns: Variable name, Presence in metadata,
  Presence in data and Concordance.
}
\description{
\code{qc_env_vars_presence} function checks if environmental variables stated
in metadata (environmental_md sheet) are really present in the environmental
data.
}
\details{
The function get all the names of the possible environmental variables and
check the presence in metadata and data. If there is any inconsistency
between them.
}
\seealso{
Other Quality Checks Functions: \code{\link{create_dic}},
  \code{\link{qc_factor_values}},
  \code{\link{qc_get_pl_md}},
  \code{\link{qc_is_timestamp}}, \code{\link{qc_md_cols}},
  \code{\link{qc_pl_treatments}},
  \code{\link{qc_sapw_area_calculator}},
  \code{\link{qc_sapw_conversion}},
  \code{\link{qc_species_verification}},
  \code{\link{qc_timestamp_errors}}
}

