% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_md_general_checks.R
\name{qc_factor_values}
\alias{qc_factor_values}
\title{Metadata factor variables check}
\usage{
qc_factor_values(site = NULL, stand = NULL, species = NULL,
  plant = NULL, environmental = NULL)
}
\arguments{
\item{site, stand, species, plant, environmental}{Data frames with the metadata
to check.}
}
\value{
A data frame with variable names, check result and NA presence as
  columns, as well as a column (Metadata) indicating the kind of metadata
  variables checked.
}
\description{
\code{qc_factor_values} function checks in the provided metadata if the
factor variables value is a valid value.
}
\details{
Values for factor variables in the metadata must be checked in order to
ensure that they are valid (i.e. they are one of the established factor
value).
}
\seealso{
Other Quality Checks Functions: \code{\link{create_dic}},
  \code{\link{qc_env_vars_presence}},
  \code{\link{qc_is_timestamp}}, \code{\link{qc_md_cols}},
  \code{\link{qc_pl_treatments}},
  \code{\link{qc_species_verification}},
  \code{\link{qc_timestamp_errors}}
}

