% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_data_general_checks.R
\name{qc_fix_timestamp}
\alias{qc_fix_timestamp}
\title{Fix TIMESTAMP formats}
\usage{
qc_fix_timestamp(data, parent_logger = "test")
}
\arguments{
\item{data}{Data frame contining the TIMESTAMP variable. Also it can be a
vector with the TIMESTAMPS values.}
}
\value{
An object of the same type of input (data frame or vector) with the
  fixed values of TIMESTAMP. If TIMESTAMP is already in format, a message
  appears and none fix is made, returning data as entered.
  If TIMESTAMP can not be fixed, an error is raised.
}
\description{
Wrapper for \code{\link{qc_is_timestamp}} and \code{\link{qc_as_timestamp}}
}
\details{
This function uses \code{\link{qc_is_timestamp}} and
\code{\link{qc_as_timestamp}} internally to check if the format is correct
and if not, try one of the known fixes to convert it in the correct
TIMESTAMP
}
\seealso{
Other Quality Checks Functions: \code{\link{create_dic}},
  \code{\link{qc_coordinates}},
  \code{\link{qc_data_results_table}},
  \code{\link{qc_env_vars_presence}},
  \code{\link{qc_factor_values}},
  \code{\link{qc_get_sapw_md}},
  \code{\link{qc_get_timestep}},
  \code{\link{qc_is_timestamp}}, \code{\link{qc_md_cols}},
  \code{\link{qc_md_results_table}},
  \code{\link{qc_pl_treatments}},
  \code{\link{qc_sapw_area_calculator}},
  \code{\link{qc_sapw_conversion}},
  \code{\link{qc_species_names_info}},
  \code{\link{qc_species_names}},
  \code{\link{qc_species_verification}},
  \code{\link{qc_timestamp_errors}}
}

