% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_coordinates_functions.R
\name{qc_get_biome}
\alias{qc_get_biome}
\title{Get the biome, temperature and precipitation of a site}
\usage{
qc_get_biome(data, merge_deserts = FALSE, parent_logger = "test")
}
\arguments{
\item{data}{Data frame of site metadata, including latitude (si_lat)
and longitude (si_long) columns that are used to obtain climatic data.}

\item{merge_deserts}{Logical indicating if desert biomes should be merged
in a single biome. By default, deserts are not merged.}
}
\value{
The same data frame with extra columns of mean annual temperature
   (si_mat), mean annual precipitation (si_map) and biome (si_biome)
}
\description{
This function takes a data frame of site metadata, including latitude (si_lat)
and longitude (si_long) columns, gets climatic data from WorldClim 1.4 and
returns the same data frame with extra columns of mean annual temperature
(si_mat), mean annual precipitation (si_map) and biome (si_biome) according to
\code{\link{qc_get_biomes_spdf}}.
}
\seealso{
Other Quality Checks Functions: \code{\link{create_dic}},
  \code{\link{qc_coordinates}},
  \code{\link{qc_data_results_table}},
  \code{\link{qc_env_ranges}},
  \code{\link{qc_env_vars_presence}},
  \code{\link{qc_ext_radiation}},
  \code{\link{qc_factor_values}},
  \code{\link{qc_fix_timestamp}},
  \code{\link{qc_get_biomes_spdf}},
  \code{\link{qc_get_sapw_md}},
  \code{\link{qc_get_timestep}},
  \code{\link{qc_get_timezone}},
  \code{\link{qc_is_timestamp}}, \code{\link{qc_md_cols}},
  \code{\link{qc_md_results_table}},
  \code{\link{qc_mind_the_gap_eff}},
  \code{\link{qc_mind_the_gap}},
  \code{\link{qc_out_hampel_filter}},
  \code{\link{qc_out_of_range}},
  \code{\link{qc_out_remove}},
  \code{\link{qc_outliers_subs}},
  \code{\link{qc_pl_treatments}},
  \code{\link{qc_rad_conversion}},
  \code{\link{qc_range_dic}}, \code{\link{qc_sapf_ranges}},
  \code{\link{qc_sapw_area_calculator}},
  \code{\link{qc_sapw_conversion}},
  \code{\link{qc_set_timezone}},
  \code{\link{qc_soil_texture}},
  \code{\link{qc_species_names_info}},
  \code{\link{qc_species_names}},
  \code{\link{qc_species_verification}},
  \code{\link{qc_start_process}},
  \code{\link{qc_time_interval}},
  \code{\link{qc_timestamp_concordance}},
  \code{\link{qc_timestamp_errors}},
  \code{\link{qc_timestamp_nas}}
}
