% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_unit_transformations.R
\name{qc_get_sapw_md}
\alias{qc_get_sapw_md}
\title{Helper function for unit conversion}
\usage{
qc_get_sapw_md(pl_metadata, parent_logger = "test")
}
\arguments{
\item{pl_metadata}{Data frame containing the plant metadata}
}
\value{
A data frame with extracted plant metadata variables as columns and
  individual plants as rows. Also a new variable \code{pl_sapw_area_est} is
  created as an empty numeric vector.
}
\description{
\code{qc_get_sapw_md} allows to extract all necessary variables from plant
metadata in order to be able to convert sap flow units.
}
\seealso{
Other Quality Checks Functions: \code{\link{create_dic}},
  \code{\link{qc_coordinates}},
  \code{\link{qc_data_results_table}},
  \code{\link{qc_env_vars_presence}},
  \code{\link{qc_factor_values}},
  \code{\link{qc_fix_timestamp}},
  \code{\link{qc_get_timestep}},
  \code{\link{qc_is_timestamp}}, \code{\link{qc_md_cols}},
  \code{\link{qc_md_results_table}},
  \code{\link{qc_mind_the_gap}},
  \code{\link{qc_pl_treatments}},
  \code{\link{qc_plot_the_gap}},
  \code{\link{qc_sapw_area_calculator}},
  \code{\link{qc_sapw_conversion}},
  \code{\link{qc_species_names_info}},
  \code{\link{qc_species_names}},
  \code{\link{qc_species_verification}},
  \code{\link{qc_time_interval}},
  \code{\link{qc_timestamp_concordance}},
  \code{\link{qc_timestamp_errors}}
}

