% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_data_general_checks.R
\name{qc_is_timestamp}
\alias{qc_is_timestamp}
\title{TimeStamp format check}
\usage{
qc_is_timestamp(data)
}
\arguments{
\item{data}{Data frame containing the data (sapflow or environmental). It
must have a \code{TIMESTAMP} variable.}
}
\value{
A message indicating if TIMESTAMP is correct or not
}
\description{
Function to check if the format of the timestamp is the correct one
}
\details{
Template timestamp is requiered to comply POSIXct format. This function
checks if the contributor followed the mandatory format.
}
\seealso{
Other Quality Checks Functions: \code{\link{create_dic}},
  \code{\link{qc_env_vars_presence}},
  \code{\link{qc_factor_values}}, \code{\link{qc_md_cols}},
  \code{\link{qc_pl_treatments}},
  \code{\link{qc_species_verification}},
  \code{\link{qc_timestamp_errors}}
}

