% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_data_general_checks.R
\name{qc_is_timestamp}
\alias{qc_is_timestamp}
\title{TimeStamp format check}
\usage{
qc_is_timestamp(data, verbose = TRUE, parent_logger = "test")
}
\arguments{
\item{data}{Data frame containing the data (sapflow or environmental). Also
a vector with the TIMESTAMP values. If data frame, \bold{it must contain} a
TIMESTAMP variable}

\item{verbose}{Logical indicating if messages of success and warnings of
failures must be presented. Default is \code{TRUE}. In order to use
inside another function, is recommended to set \code{verbose = FALSE}.}
}
\value{
A message/warning indicating if TIMESTAMP is correct or not. Also
  an invisible logical object is returned, indicating success (TRUE) or
  failure (FALSE).
}
\description{
Function to check if the format of the timestamp is the correct one
}
\details{
Template timestamp is requiered to comply POSIXct format. This function
checks if the contributor followed the mandatory format.
}
\seealso{
Other Quality Checks Functions: \code{\link{create_dic}},
  \code{\link{qc_coordinates}},
  \code{\link{qc_data_results_table}},
  \code{\link{qc_env_vars_presence}},
  \code{\link{qc_factor_values}},
  \code{\link{qc_fix_timestamp}},
  \code{\link{qc_get_sapw_md}},
  \code{\link{qc_get_timestep}},
  \code{\link{qc_get_timezone}}, \code{\link{qc_md_cols}},
  \code{\link{qc_md_results_table}},
  \code{\link{qc_mind_the_gap}},
  \code{\link{qc_pl_treatments}},
  \code{\link{qc_sapw_area_calculator}},
  \code{\link{qc_sapw_conversion}},
  \code{\link{qc_set_timezone}},
  \code{\link{qc_species_names_info}},
  \code{\link{qc_species_names}},
  \code{\link{qc_species_verification}},
  \code{\link{qc_time_interval}},
  \code{\link{qc_timestamp_concordance}},
  \code{\link{qc_timestamp_errors}}
}

