% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_md_general_checks.R
\name{qc_md_cols}
\alias{qc_md_cols}
\title{Metadata columns check}
\usage{
qc_md_cols(metadata, dic, parent_logger = "test")
}
\arguments{
\item{metadata}{Data frame containing the metadata in which the test will be
made.}

\item{dic}{Name of the metadata dictionary to use as character. It must be
one of the following: \code{'stand_md'}, \code{'site_md'},
\code{'species_md'}, \code{'plant_md'} or \code{'environmental_md'}}
}
\value{
A data frame with variable names in one column and result of the
  checks in teh following columns
}
\description{
\code{qc_md_cols} checks if the columns of the provided metadata return the
correct class, as well as indicates any NAs present and the absence of any
mandatory variable.
}
\details{
After loading the metadata, columns classes and presence/absence of mandatory
variables must be checked in order to continue with the quality check
workflow. This function returns a summary of the metadata columns and their
state.
}
\seealso{
Other Quality Checks Functions: \code{\link{create_dic}},
  \code{\link{qc_coordinates}},
  \code{\link{qc_data_results_table}},
  \code{\link{qc_env_vars_presence}},
  \code{\link{qc_factor_values}},
  \code{\link{qc_fix_timestamp}},
  \code{\link{qc_get_biomes_spdf}},
  \code{\link{qc_get_biome}}, \code{\link{qc_get_sapw_md}},
  \code{\link{qc_get_timestep}},
  \code{\link{qc_get_timezone}},
  \code{\link{qc_is_timestamp}},
  \code{\link{qc_md_results_table}},
  \code{\link{qc_mind_the_gap_eff}},
  \code{\link{qc_mind_the_gap}},
  \code{\link{qc_pl_treatments}},
  \code{\link{qc_rad_conversion}},
  \code{\link{qc_sapf_range_check}},
  \code{\link{qc_sapw_area_calculator}},
  \code{\link{qc_sapw_conversion}},
  \code{\link{qc_set_timezone}},
  \code{\link{qc_soil_texture}},
  \code{\link{qc_solar_timestamp}},
  \code{\link{qc_species_names_info}},
  \code{\link{qc_species_names}},
  \code{\link{qc_species_verification}},
  \code{\link{qc_start_process}},
  \code{\link{qc_time_interval}},
  \code{\link{qc_timestamp_concordance}},
  \code{\link{qc_timestamp_errors}},
  \code{\link{qc_timestamp_nas}}
}

