% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_gaps_functions.R
\name{qc_mind_the_gap}
\alias{qc_mind_the_gap}
\title{Gaps info}
\usage{
qc_mind_the_gap(data, trim = FALSE, parent_logger = "test")
}
\arguments{
\item{data}{Data frame containing the data in which obtain info about gaps.
It must have a TIMESTAMP variable}

\item{trim}{Logial indicating if starting and ending gaps must be included.
Default to FALSE.}
}
\value{
A data frame with info about gaps:
  \itemize{
      \item gap_start: Gap start TIMESTAMP
      \item gap_end: Gap end TIMESTAMP
      \item gap_interval: Gap interval in minutes
      \item gap_coverage: Gap coverage in percentage
      \item timestamp_start: Absolute timestamp start
      \item timestamp_end: Absolute timestamp end
  }
}
\description{
Information about gaps
}
\seealso{
Other Quality Checks Functions: \code{\link{create_dic}},
  \code{\link{qc_coordinates}},
  \code{\link{qc_data_results_table}},
  \code{\link{qc_env_vars_presence}},
  \code{\link{qc_factor_values}},
  \code{\link{qc_fix_timestamp}},
  \code{\link{qc_get_sapw_md}},
  \code{\link{qc_get_timestep}},
  \code{\link{qc_is_timestamp}}, \code{\link{qc_md_cols}},
  \code{\link{qc_md_results_table}},
  \code{\link{qc_pl_treatments}},
  \code{\link{qc_sapw_area_calculator}},
  \code{\link{qc_sapw_conversion}},
  \code{\link{qc_species_names_info}},
  \code{\link{qc_species_names}},
  \code{\link{qc_species_verification}},
  \code{\link{qc_time_interval}},
  \code{\link{qc_timestamp_concordance}},
  \code{\link{qc_timestamp_errors}}
}

