% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_outliers_and_ranges.R
\name{qc_out_remove}
\alias{qc_out_remove}
\title{Outliers QC}
\usage{
qc_out_remove(
  sfn_data,
  k = 25L,
  t0 = 10L,
  method = "hampel",
  reverse = TRUE,
  substitute = FALSE,
  parent_logger = "test"
)
}
\arguments{
\item{sfn_data}{SfnData object with the site data and metadata}

\item{k}{window semi-value (integer) for substitution value calculation. See
\code{\link{out_hampel_filter}} for details.}

\item{t0}{Integer value indicating the number of standard deviations for
outlier detection threshold.}

\item{method}{Character vector indicating the method to use in the outlier
estimation: \code{hampel} (default), \code{tukey} and \code{quantile}.}

\item{reverse}{Logical indicating if in case of k is bigger than the data
length (i.e in the extremes of the vector) a reverse replication must be
done. Default to TRUE.}

\item{substitute}{Logical indicating if the outlier subsitution must be made.
Default to FALSE.}
}
\value{
a SfnData object as the one provided with the oulier values
substituted and the flags slot updated.
}
\description{
Outliers detection, substitution and annotation for SfnData objects
}
\details{
Outliers for sap flow data and environmental data are located and substituted
by the selected method, using the \code{\link{out_hampel_filter}} function.
}
\seealso{
Other Quality Checks Functions: 
\code{\link{create_dic}()},
\code{\link{qc_coordinates}()},
\code{\link{qc_data_results_table}()},
\code{\link{qc_env_ranges}()},
\code{\link{qc_env_vars_presence}()},
\code{\link{qc_ext_radiation}()},
\code{\link{qc_factor_values}()},
\code{\link{qc_fix_timestamp}()},
\code{\link{qc_get_biome}()},
\code{\link{qc_get_biomes_spdf}()},
\code{\link{qc_get_sapw_md}()},
\code{\link{qc_get_timestep}()},
\code{\link{qc_get_timezone}()},
\code{\link{qc_is_timestamp}()},
\code{\link{qc_md_cols}()},
\code{\link{qc_md_results_table}()},
\code{\link{qc_mind_the_gap}()},
\code{\link{qc_mind_the_gap_deprecated}()},
\code{\link{qc_mind_the_gap_eff}()},
\code{\link{qc_out_hampel_filter}()},
\code{\link{qc_out_of_range}()},
\code{\link{qc_outliers_process}()},
\code{\link{qc_outliers_subs}},
\code{\link{qc_pl_treatments}()},
\code{\link{qc_rad_conversion}()},
\code{\link{qc_range_dic}()},
\code{\link{qc_sapf_ranges}()},
\code{\link{qc_sapw_area_calculator}()},
\code{\link{qc_sapw_conversion}()},
\code{\link{qc_set_timezone}()},
\code{\link{qc_soil_texture}()},
\code{\link{qc_species_names}()},
\code{\link{qc_species_names_info}()},
\code{\link{qc_species_verification}()},
\code{\link{qc_start_process}()},
\code{\link{qc_swc_check}()},
\code{\link{qc_swc_fix}()},
\code{\link{qc_time_interval}()},
\code{\link{qc_timestamp_concordance}()},
\code{\link{qc_timestamp_errors}()},
\code{\link{qc_timestamp_nas}()}
}
\concept{Quality Checks Functions}
