% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_outliers_and_ranges.R
\name{qc_outliers_subs}
\alias{qc_outliers_subs}
\alias{qc_out_tukeyline}
\alias{qc_out_medianreg}
\alias{qc_out_median}
\title{Outliers substitution}
\usage{
qc_out_tukeyline(y, k = 25L, parent_logger = "test")

qc_out_medianreg(y, k = 25L, parent_logger = "test")

qc_out_median(y, k = 25L, parent_logger = "test")
}
\arguments{
\item{y}{vector of values for outlier substitution}

\item{k}{window semi-value (integer) for substitution value calculation. See
details.}
}
\value{
a vector of the same lengh of y with the outlier values substituted
by the calculated value
}
\description{
Outlier substitution by different methods
}
\details{
\code{k} is the semi-value of the desired window for substitute value
calculation. The window is formed by the range \code{y[i] - k : y[i] + k}.
}
\section{Functions}{
\itemize{
\item \code{qc_out_tukeyline()}: 

\item \code{qc_out_medianreg()}: 

\item \code{qc_out_median()}: 

}}
\section{out_tukeyline}{

\code{out_tukeyline} performs a robust fitting with the \code{line} function
}

\section{out_medianreg}{

\code{out_medianreg} performs a regression of medians with the quantreg package
}

\section{out_median}{

\code{out_median} performs a classic median outlier detection
}

\seealso{
Other Quality Checks Functions: 
\code{\link{create_dic}()},
\code{\link{qc_coordinates}()},
\code{\link{qc_data_results_table}()},
\code{\link{qc_env_ranges}()},
\code{\link{qc_env_vars_presence}()},
\code{\link{qc_ext_radiation}()},
\code{\link{qc_factor_values}()},
\code{\link{qc_fix_timestamp}()},
\code{\link{qc_get_biome}()},
\code{\link{qc_get_biomes_spdf}()},
\code{\link{qc_get_sapw_md}()},
\code{\link{qc_get_timestep}()},
\code{\link{qc_get_timezone}()},
\code{\link{qc_is_timestamp}()},
\code{\link{qc_md_cols}()},
\code{\link{qc_md_results_table}()},
\code{\link{qc_mind_the_gap}()},
\code{\link{qc_mind_the_gap_eff}()},
\code{\link{qc_out_hampel_filter}()},
\code{\link{qc_out_of_range}()},
\code{\link{qc_out_remove}()},
\code{\link{qc_outliers_process}()},
\code{\link{qc_pl_treatments}()},
\code{\link{qc_rad_conversion}()},
\code{\link{qc_range_dic}()},
\code{\link{qc_sapf_ranges}()},
\code{\link{qc_sapw_area_calculator}()},
\code{\link{qc_sapw_conversion}()},
\code{\link{qc_set_timezone}()},
\code{\link{qc_soil_texture}()},
\code{\link{qc_species_names}()},
\code{\link{qc_species_names_info}()},
\code{\link{qc_species_verification}()},
\code{\link{qc_start_process}()},
\code{\link{qc_swc_check}()},
\code{\link{qc_swc_fix}()},
\code{\link{qc_time_interval}()},
\code{\link{qc_timestamp_concordance}()},
\code{\link{qc_timestamp_errors}()},
\code{\link{qc_timestamp_nas}()}
}
\concept{Quality Checks Functions}
