% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plant_functions.R
\name{qc_pl_treatments}
\alias{qc_pl_treatments}
\title{Check plant treatments for misspelling and concordance errors}
\usage{
qc_pl_treatments(plant_md)
}
\arguments{
\item{plant_md}{Data frame containing the plant metadata}
}
\value{
A data frame with the information about the different treatments if
  any.
}
\description{
\code{qc_pl_treatments} function looks for spelling errors in treatments
provided in the plant metadata.
}
\details{
This function recopiles all declared treatments and look for mispelling errors
in the treatments (i.e. \code{Drought} and \code{drought}) that can be cause
of later analysis errors.
}
\seealso{
Other Quality Checks Functions: \code{\link{create_dic}},
  \code{\link{qc_md_cols}},
  \code{\link{qc_species_verification}}
}

