% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_unit_transformations.R
\name{qc_rh}
\alias{qc_rh}
\title{Calculation of rh from vpd and ta}
\usage{
qc_rh(data, parent_logger = "test")
}
\arguments{
\item{data}{Data frame containing the environmental data}
}
\value{
a Data frame of the environmental data with the new rh variable
  calculated
}
\description{
Calculate the rh if needed from other environmental variables
}
\details{
The calculations for this function were obtained from "Plants and Microclimate"
by Hamlyn G. Jones (Cambridge University Press, 3rd Edition). The enhancement
factor that corrects for the slight departure of the behaviour of water in
air from that of a pure gas is NOT implemented
}
\seealso{
Other Unit conversion: \code{\link{qc_cm_cm_h}},
  \code{\link{qc_transf_list}},
  \code{\link{qc_transformation_vars}},
  \code{\link{qc_units_process}}, \code{\link{qc_vpd}}
}
