% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_unit_transformations.R
\name{qc_sapw_area_calculator}
\alias{qc_sapw_area_calculator}
\title{Sapwood area calculator}
\usage{
qc_sapw_area_calculator(pl_vars, parent_logger = "test")
}
\arguments{
\item{pl_vars}{Data frame containing the needed variables, usually the result
of \code{\link{qc_get_sapw_md}}.}
}
\value{
A data frame, exactly as returned by \code{\link{qc_get_sapw_md}},
but with the variable pl_sapw_area_est filled with the estimated values.
}
\description{
Function to calculate sapwood area from DBH, bark thickness and sapwood depth
}
\details{
In the case that sapwood area variable (\code{pl_sapw_area}) is not provided,
it still is needed to perform the unit conversion. In this case sapwood area
can be estimated from DBH and sapwood depth and bark thickness. If bark
thickness is not provided, estimation can be made but results should be
revised as, depending on the species, error in the estimates can be large.
}
\section{Ring area formula}{

The area of a ring is \eqn{\pi(R² - r²)}, where \code{R} is the radius
including the ring and \code{r} is the radius till the ring. Thus,
\eqn{R = (dbh / 2) - bark_thickness} and
\eqn{r = (dbh / 2) - bark_thickness - sapwood_depth}.
}

\seealso{
Other Quality Checks Functions: 
\code{\link{create_dic}()},
\code{\link{qc_coordinates}()},
\code{\link{qc_data_results_table}()},
\code{\link{qc_env_ranges}()},
\code{\link{qc_env_vars_presence}()},
\code{\link{qc_ext_radiation}()},
\code{\link{qc_factor_values}()},
\code{\link{qc_fix_timestamp}()},
\code{\link{qc_get_biome}()},
\code{\link{qc_get_biomes_spdf}()},
\code{\link{qc_get_sapw_md}()},
\code{\link{qc_get_timestep}()},
\code{\link{qc_get_timezone}()},
\code{\link{qc_is_timestamp}()},
\code{\link{qc_md_cols}()},
\code{\link{qc_md_results_table}()},
\code{\link{qc_mind_the_gap}()},
\code{\link{qc_mind_the_gap_deprecated}()},
\code{\link{qc_mind_the_gap_eff}()},
\code{\link{qc_out_hampel_filter}()},
\code{\link{qc_out_of_range}()},
\code{\link{qc_out_remove}()},
\code{\link{qc_outliers_process}()},
\code{\link{qc_outliers_subs}},
\code{\link{qc_pl_treatments}()},
\code{\link{qc_rad_conversion}()},
\code{\link{qc_range_dic}()},
\code{\link{qc_sapf_ranges}()},
\code{\link{qc_sapw_conversion}()},
\code{\link{qc_set_timezone}()},
\code{\link{qc_soil_texture}()},
\code{\link{qc_species_names}()},
\code{\link{qc_species_names_info}()},
\code{\link{qc_species_verification}()},
\code{\link{qc_start_process}()},
\code{\link{qc_swc_check}()},
\code{\link{qc_swc_fix}()},
\code{\link{qc_time_interval}()},
\code{\link{qc_timestamp_concordance}()},
\code{\link{qc_timestamp_errors}()},
\code{\link{qc_timestamp_nas}()}
}
\concept{Quality Checks Functions}
