% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_unit_transformations.R
\name{qc_sapw_conversion}
\alias{qc_sapw_conversion}
\title{Sap flow units transformation}
\usage{
qc_sapw_conversion(
  data,
  sapw_md,
  output_units = "plant",
  parent_logger = "test"
)
}
\arguments{
\item{data}{Data frame containing the sap flow measurements}

\item{sapw_md}{Data frame containing the sapwood metadata, as obtained from
\code{\link{qc_get_sapw_md}} or \code{\link{qc_sapw_area_calculator}}.}

\item{output_units}{Character vector indicating the kind of output units.
Allowed values are \code{'plant'}, \code{'sapwood'} and \code{'leaf'}.
See details to obtain more information}
}
\description{
Function to transform between sap flow units
}
\details{
Sap flow accepted units can be of two kinds, \code{per sapwood area} and
\code{per plant}. Both kinds can come in many flavours and unit conversion
must be done to allow data integration and analysis.
This function can return three kind of units:
\describe{
\item{\bold{Per sapwood  area}}{
In this case, units returned are \eqn{cm³·cm⁻²·h⁻¹}
  }
  \item{\bold{Per plant}}{
In this case, units returned are \eqn{cm³·h⁻¹}
  }
  \item{\bold{Per leaf area unit}}{
In this case, units returned are \eqn{cm³·cm⁻²·h⁻¹}
  }
}
}
\section{Sapwood area}{

If origin units are \emph{per sapwood area}, direct transformation is made. If
origin units are \emph{per plant} then \code{pl_sapw_area} variable from
plant metadata is needed to make the conversion.
}

\section{Plant}{

If origin units are \emph{per plant}, direct transformation is made. If
origin units are \emph{per sapwood area} then \code{pl_sapw_area} variable
from plant metadata is needed to make the conversion.
}

\section{Leaf area}{

If origin units are \emph{per plant} then \code{pl_leaf_area} variable from
plant metadata is needed to make the conversion. If origin units are
\emph{per sapwood area} then \code{pl_leaf_area} and \code{pl_sapw_area}
variables from plant metadata are needed.
}

\section{\code{pl_sapw_area}}{

If \code{pl_sapw_area} is not available but \code{pl_sapw_depth} and
\code{pl_dbh} are provided, sapwood area value can be estimated by means of
\code{\link{qc_sapw_area_calculator}} function previous to the use of this
function.
}

\seealso{
Other Quality Checks Functions: 
\code{\link{create_dic}()},
\code{\link{qc_coordinates}()},
\code{\link{qc_data_results_table}()},
\code{\link{qc_env_ranges}()},
\code{\link{qc_env_vars_presence}()},
\code{\link{qc_ext_radiation}()},
\code{\link{qc_factor_values}()},
\code{\link{qc_fix_timestamp}()},
\code{\link{qc_get_biome}()},
\code{\link{qc_get_biomes_spdf}()},
\code{\link{qc_get_sapw_md}()},
\code{\link{qc_get_timestep}()},
\code{\link{qc_get_timezone}()},
\code{\link{qc_is_timestamp}()},
\code{\link{qc_md_cols}()},
\code{\link{qc_md_results_table}()},
\code{\link{qc_mind_the_gap}()},
\code{\link{qc_mind_the_gap_eff}()},
\code{\link{qc_out_hampel_filter}()},
\code{\link{qc_out_of_range}()},
\code{\link{qc_out_remove}()},
\code{\link{qc_outliers_process}()},
\code{\link{qc_outliers_subs}},
\code{\link{qc_pl_treatments}()},
\code{\link{qc_rad_conversion}()},
\code{\link{qc_range_dic}()},
\code{\link{qc_sapf_ranges}()},
\code{\link{qc_sapw_area_calculator}()},
\code{\link{qc_set_timezone}()},
\code{\link{qc_soil_texture}()},
\code{\link{qc_species_names}()},
\code{\link{qc_species_names_info}()},
\code{\link{qc_species_verification}()},
\code{\link{qc_start_process}()},
\code{\link{qc_swc_check}()},
\code{\link{qc_swc_fix}()},
\code{\link{qc_time_interval}()},
\code{\link{qc_timestamp_concordance}()},
\code{\link{qc_timestamp_errors}()},
\code{\link{qc_timestamp_nas}()}
}
\concept{Quality Checks Functions}
