% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_species_functions.R
\name{qc_species_names}
\alias{qc_species_names}
\title{Wrapper for species names check}
\usage{
qc_species_names(species, conservatism = 0.9, parent_logger = "test")
}
\arguments{
\item{species}{Character vector containing the species names to verify}

\item{conservatism}{Numerical value between 0 and 1 indicating the
conservatism level of the tpl spelling check algorithm. Default to
0.9}
}
\value{
a vector with the fixed names of the species if fix is possible or
  needed, or a vector with the original names of the species if the fix is not
  possible or they are correct
}
\description{
Wrapper for \code{\link{qc_species_names_info}} and \code{\link{qc_species_names_fix}}
}
\seealso{
Other Quality Checks Functions: \code{\link{create_dic}},
  \code{\link{qc_coordinates}},
  \code{\link{qc_data_results_table}},
  \code{\link{qc_env_vars_presence}},
  \code{\link{qc_factor_values}},
  \code{\link{qc_fix_timestamp}},
  \code{\link{qc_get_sapw_md}},
  \code{\link{qc_get_timestep}},
  \code{\link{qc_is_timestamp}}, \code{\link{qc_md_cols}},
  \code{\link{qc_md_results_table}},
  \code{\link{qc_pl_treatments}},
  \code{\link{qc_sapw_area_calculator}},
  \code{\link{qc_sapw_conversion}},
  \code{\link{qc_species_names_info}},
  \code{\link{qc_species_verification}},
  \code{\link{qc_timestamp_errors}}
}

