% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_species_functions.R
\name{qc_species_names}
\alias{qc_species_names}
\title{Verify provided species names (spelling and correctness)}
\usage{
qc_species_names(species, conservatism = 0.9)
}
\arguments{
\item{species}{Character vector containing the species names to verify}

\item{conservatism}{Numerical value between 0 and 1 indicating the
conservatism level of the tpl spelling check algorithm. Default to
0.75}
}
\value{
A character vector with species fixed in spelling and correctness.
}
\description{
\code{qc_species_names} uses \code{tpl} package
(\code{\link{https://github.com/gustavobio/tpl}}) to verify the species
introduced by the contributors and fix spelling errors.
}
\details{
This function takes a vector of species names and check if they are right
spelled. Also, if a synonym is used, the function changes it automatically in
order to have the same name for the same species.
}
\section{GitHub package}{

\code{tpl} and \code{tpldata} are GitHub packages, and they are not available
for install in the ususal way. In order to achieve that
\code{qc_species_verification} works as expected, manual installation of
those two packages must be done previously:
\code{devtools::install_github("gustavobio/tpldata")}
\code{devtools::install_github("gustavobio/tpl")}
}

