% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_species_functions.R
\name{qc_species_names_info}
\alias{qc_species_names_info}
\title{Info of species names spelling}
\usage{
qc_species_names_info(species, conservatism = 0.9, parent_logger = "test")
}
\arguments{
\item{species}{Character vector containing the species names to verify}

\item{conservatism}{Numerical value between 0 and 1 indicating the
conservatism level of the tpl spelling check algorithm. Default to
0.9}
}
\value{
A data frame summarizing the species names declared, the species
  names obtained after tpl and the concordance and NAs info
}
\description{
Summary of species names spelling info
}
\seealso{
Other Quality Checks Functions: \code{\link{create_dic}},
  \code{\link{qc_coordinates}},
  \code{\link{qc_data_results_table}},
  \code{\link{qc_env_vars_presence}},
  \code{\link{qc_factor_values}},
  \code{\link{qc_fix_timestamp}},
  \code{\link{qc_get_sapw_md}},
  \code{\link{qc_get_timestep}},
  \code{\link{qc_get_timezone}},
  \code{\link{qc_is_timestamp}}, \code{\link{qc_md_cols}},
  \code{\link{qc_md_results_table}},
  \code{\link{qc_mind_the_gap_eff}},
  \code{\link{qc_mind_the_gap}},
  \code{\link{qc_pl_treatments}},
  \code{\link{qc_sapf_range_check}},
  \code{\link{qc_sapw_area_calculator}},
  \code{\link{qc_sapw_conversion}},
  \code{\link{qc_set_timezone}},
  \code{\link{qc_solar_timestamp}},
  \code{\link{qc_species_names}},
  \code{\link{qc_species_verification}},
  \code{\link{qc_start_process}},
  \code{\link{qc_time_interval}},
  \code{\link{qc_timestamp_concordance}},
  \code{\link{qc_timestamp_errors}},
  \code{\link{qc_timestamp_nas}}
}

