% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_species_functions.R
\name{qc_species_verification}
\alias{qc_species_verification}
\title{Check if sp_name and sp_ntrees coincides with plants specified in plant_md}
\usage{
qc_species_verification(species_md, plant_md)
}
\arguments{
\item{species_md}{Data frame containing species metadata.}

\item{plant_md}{Data frame containing plant metadata.}
}
\value{
No idea for the moment
}
\description{
\code{qc_species_verification} checks for coincidence in the number and species
names between species_md and plant_md
}
\details{
In order to check if provided species metadata coincides with provided plant
metadata, species names and number of plants in each species are checked
against plant metadata.
}
\seealso{
Other Quality Checks Functions: \code{\link{create_dic}},
  \code{\link{qc_md_cols}}
}

