% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_species_functions.R
\name{qc_species_verification}
\alias{qc_species_verification}
\title{Check if sp_name and sp_ntrees coincides with plants specified in plant_md}
\usage{
qc_species_verification(species_md, plant_md, parent_logger = "test")
}
\arguments{
\item{species_md}{Data frame containing species metadata.}

\item{plant_md}{Data frame containing plant metadata.}
}
\value{
A data frame with the species provided in plant and species metadata,
  as well as the number of trees of each species in both metadata.
  \code{coincidence} column indicates result of checking both numbers and
  presence of the same species.
}
\description{
\code{qc_species_verification} checks for coincidence in the number and species
names between species_md and plant_md
}
\details{
In order to check if provided species metadata coincides with provided plant
metadata, species names and number of plants in each species are checked
against plant metadata.
}
\section{Coincidence}{

  Possible values of \code{coincidence} column in results data frame are:
  \code{TRUE}, indicating both metadata have the same species and the number
  of trees are the same; \code{FALSE} indicates that both metadata have the
  same species, but the number of trees are no the same; finally,
  \code{NA} indicates that the species are not the same in both metadata.
}
\seealso{
Other Quality Checks Functions: \code{\link{create_dic}},
  \code{\link{qc_coordinates}},
  \code{\link{qc_data_results_table}},
  \code{\link{qc_env_vars_presence}},
  \code{\link{qc_ext_radiation}},
  \code{\link{qc_factor_values}},
  \code{\link{qc_fix_timestamp}},
  \code{\link{qc_get_biomes_spdf}},
  \code{\link{qc_get_biome}}, \code{\link{qc_get_sapw_md}},
  \code{\link{qc_get_timestep}},
  \code{\link{qc_get_timezone}},
  \code{\link{qc_is_timestamp}}, \code{\link{qc_md_cols}},
  \code{\link{qc_md_results_table}},
  \code{\link{qc_mind_the_gap_eff}},
  \code{\link{qc_mind_the_gap}},
  \code{\link{qc_pl_treatments}},
  \code{\link{qc_rad_conversion}},
  \code{\link{qc_range_dic}},
  \code{\link{qc_sapw_area_calculator}},
  \code{\link{qc_sapw_conversion}},
  \code{\link{qc_set_timezone}},
  \code{\link{qc_soil_texture}},
  \code{\link{qc_species_names_info}},
  \code{\link{qc_species_names}},
  \code{\link{qc_start_process}},
  \code{\link{qc_time_interval}},
  \code{\link{qc_timestamp_concordance}},
  \code{\link{qc_timestamp_errors}},
  \code{\link{qc_timestamp_nas}}
}

