% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_main.R
\name{qc_start_process}
\alias{qc_start_process}
\title{Start QC process}
\usage{
qc_start_process(folder = ".", rdata = TRUE, parent_logger = "test")
}
\arguments{
\item{folder}{Character string with the route to the folder to start QC
process}

\item{rdata}{Logical indicating if objects created in the QC must be saved in
a file}
}
\value{
Invisible TRUE if all the process is ok, and invisible FALSE if there
  was some error in the process
}
\description{
Start QC process from the site code.
}
\details{
This function check the status of the site, starts it if it does not exists,
and start the QC process
}
\seealso{
Other Quality Checks Functions: \code{\link{create_dic}},
  \code{\link{qc_coordinates}},
  \code{\link{qc_data_results_table}},
  \code{\link{qc_env_vars_presence}},
  \code{\link{qc_factor_values}},
  \code{\link{qc_fix_timestamp}},
  \code{\link{qc_get_biomes_spdf}},
  \code{\link{qc_get_biome}}, \code{\link{qc_get_sapw_md}},
  \code{\link{qc_get_timestep}},
  \code{\link{qc_get_timezone}},
  \code{\link{qc_is_timestamp}}, \code{\link{qc_md_cols}},
  \code{\link{qc_md_results_table}},
  \code{\link{qc_mind_the_gap_eff}},
  \code{\link{qc_mind_the_gap}},
  \code{\link{qc_pl_treatments}},
  \code{\link{qc_rad_conversion}},
  \code{\link{qc_sapf_range_check}},
  \code{\link{qc_sapw_area_calculator}},
  \code{\link{qc_sapw_conversion}},
  \code{\link{qc_set_timezone}},
  \code{\link{qc_solar_timestamp}},
  \code{\link{qc_species_names_info}},
  \code{\link{qc_species_names}},
  \code{\link{qc_species_verification}},
  \code{\link{qc_time_interval}},
  \code{\link{qc_timestamp_concordance}},
  \code{\link{qc_timestamp_errors}},
  \code{\link{qc_timestamp_nas}}
}

