% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_data_general_checks.R
\name{qc_timestamp_concordance}
\alias{qc_timestamp_concordance}
\title{Check concordance between sapflow TIMESTAMP and env TIMESTAMP}
\usage{
qc_timestamp_concordance(sapf_data = NULL, env_data = NULL,
  sapf_intervals = NULL, env_intervals = NULL, plot = FALSE,
  parent_logger = "test")
}
\arguments{
\item{sapf_data}{Data frame containing the sapflow data and its TIMESTAMP}

\item{env_data}{Data frame containing the environmental data and its
TIMESTAMP}

\item{sapf_intervals}{Data frame obtained from \code{\link{qc_time_interval}},
optional. It can be used if no sapf and env data is provided}

\item{env_intervals}{Data frame obtained from \code{\link{qc_time_interval}},
optional. It can be used if no sapf and env data is provided}

\item{plot}{Logical indicating if the result is presented in graphical mode
(ggplot2 object). Desfault to TRUE.}
}
\value{
A data frame summarising the results
}
\description{
Are the sapflow and environmental TIMESTAMPS in concordance?
}
\details{
This function uses \code{\link{qc_time_interval}} internally to perform
operations with time intervals. There is two ways of using this function:
\enumerate{
  \item Directly from sapflow and environmental data, providing both datasets
  \item From the results obtained from \code{\link{qc_time_interval}},
        providing both data frames with the results.
}
}
\seealso{
Other Quality Checks Functions: \code{\link{create_dic}},
  \code{\link{qc_coordinates}},
  \code{\link{qc_data_results_table}},
  \code{\link{qc_env_ranges}},
  \code{\link{qc_env_vars_presence}},
  \code{\link{qc_ext_radiation}},
  \code{\link{qc_factor_values}},
  \code{\link{qc_fix_timestamp}},
  \code{\link{qc_get_biomes_spdf}},
  \code{\link{qc_get_biome}}, \code{\link{qc_get_sapw_md}},
  \code{\link{qc_get_timestep}},
  \code{\link{qc_get_timezone}},
  \code{\link{qc_is_timestamp}}, \code{\link{qc_md_cols}},
  \code{\link{qc_md_results_table}},
  \code{\link{qc_mind_the_gap_eff}},
  \code{\link{qc_mind_the_gap}},
  \code{\link{qc_out_hampel_filter}},
  \code{\link{qc_out_of_range}},
  \code{\link{qc_out_remove}},
  \code{\link{qc_outliers_process}},
  \code{\link{qc_outliers_subs}},
  \code{\link{qc_pl_treatments}},
  \code{\link{qc_rad_conversion}},
  \code{\link{qc_range_dic}}, \code{\link{qc_sapf_ranges}},
  \code{\link{qc_sapw_area_calculator}},
  \code{\link{qc_sapw_conversion}},
  \code{\link{qc_set_timezone}},
  \code{\link{qc_soil_texture}},
  \code{\link{qc_species_names_info}},
  \code{\link{qc_species_names}},
  \code{\link{qc_species_verification}},
  \code{\link{qc_start_process}},
  \code{\link{qc_time_interval}},
  \code{\link{qc_timestamp_errors}},
  \code{\link{qc_timestamp_nas}}
}
