% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_data_general_checks.R
\name{qc_timestamp_errors}
\alias{qc_timestamp_errors}
\title{TimeStamp errors localization}
\usage{
qc_timestamp_errors(data, timestep = 15, parent_logger = "test")
}
\arguments{
\item{data}{Data frame containing the data (sapflow or environmental). It
must have a \code{TIMESTAMP} variable.}

\item{timestep}{Integer indicating the measures timestep (in minutes).}
}
\value{
A data frame summarizing the errors found, indicating the interval
  and its duration (in seconds).
}
\description{
Function to pinpoint errors in the timestamp
}
\details{
\code{TIMESTAMP} variable can present continuity, date or another kind of
errors. This function checks for them. For that, from the timestep declared
in the metadata, a summary of intervals differing from it is
presented. Only intervals differing more than 59 seconds from the declared
timestep are reported.
}
\seealso{
Other Quality Checks Functions: \code{\link{create_dic}},
  \code{\link{qc_coordinates}},
  \code{\link{qc_data_results_table}},
  \code{\link{qc_env_vars_presence}},
  \code{\link{qc_ext_radiation}},
  \code{\link{qc_factor_values}},
  \code{\link{qc_fix_timestamp}},
  \code{\link{qc_get_biomes_spdf}},
  \code{\link{qc_get_biome}}, \code{\link{qc_get_sapw_md}},
  \code{\link{qc_get_timestep}},
  \code{\link{qc_get_timezone}},
  \code{\link{qc_is_timestamp}}, \code{\link{qc_md_cols}},
  \code{\link{qc_md_results_table}},
  \code{\link{qc_mind_the_gap_eff}},
  \code{\link{qc_mind_the_gap}},
  \code{\link{qc_pl_treatments}},
  \code{\link{qc_rad_conversion}},
  \code{\link{qc_sapf_range_check}},
  \code{\link{qc_sapw_area_calculator}},
  \code{\link{qc_sapw_conversion}},
  \code{\link{qc_set_timezone}},
  \code{\link{qc_soil_texture}},
  \code{\link{qc_species_names_info}},
  \code{\link{qc_species_names}},
  \code{\link{qc_species_verification}},
  \code{\link{qc_start_process}},
  \code{\link{qc_time_interval}},
  \code{\link{qc_timestamp_concordance}},
  \code{\link{qc_timestamp_nas}}
}

