% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_unit_transformations.R
\name{qc_transformation_vars}
\alias{qc_transformation_vars}
\title{Presence of variables needed for transformations summary}
\usage{
qc_transformation_vars(sfndata, parent_logger = "test")
}
\arguments{
\item{sfndata}{SfnData object for the site containing all the information}
}
\value{
A data frame with the following columns:

  \itemize{
    \item{Variable: Variable name}
    \item{Location: Variable location (i.e. env_data or env_md)}
    \item{Trasformation: Tranformation/Conversion for whihc the variable is needed}
    \item{Presence: Logical indicating if the variable is present}
  }
}
\description{
Summary table for variables needed for unit and any other kind of conversions
}
\details{
This function generates a table (data frame) with information about the
presence of the variables needed for unit, radiation, solar time and others
data transformations/conversions.
}
\seealso{
Other Unit conversion: \code{\link{qc_cm_cm_h}},
  \code{\link{qc_transf_list}}
}

