% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{sfn_data_constructor}
\alias{sfn_data_constructor}
\title{Create a SfnData object from results of Quality Checks}
\usage{
sfn_data_constructor(sapf_data = NULL, env_data = NULL, site_md = NULL,
  stand_md = NULL, species_md = NULL, plant_md = NULL, env_md = NULL,
  parent_logger = "test")
}
\arguments{
\item{sapf_data}{Data frame with sapflow data after QC process}

\item{env_data}{Data frame with environmental data after QC process}

\item{site_md}{Data frame with the site metadata after QC process}

\item{stand_md}{Data frame with the stand metadata after QC process}

\item{species_md}{Data frame with the species metadata after QC process}

\item{plant_md}{Data frame with the plant metadata after QC process}

\item{env_md}{Data frame with the environmental after QC process}
}
\value{
A SfnData object with all the data and metadata of the site
}
\description{
This function gather all the data and metadata generated in the QC process
and build an SfnData class object
}
\details{
\code{sfn_data_constructor} function generates a SfnData object containing
all relevant data and metadata for a site. Sapflow and environmental data
are converted to the same timestamp span and added rows are flagged as
\code{NA_ADDED}. Original NAs are also flagged as \code{NA_PRESENT}. For
info about the available slots see \code{\link{SfnData}},
\code{\link{sfn_get_methods}} and \code{\link{sfn_replacement}}
}
\seealso{
Other Data Flow: \code{\link{df_accepted_to_lvl1}},
  \code{\link{df_copy_templates}},
  \code{\link{df_folder_structure}},
  \code{\link{df_get_data_folders}},
  \code{\link{df_get_status}},
  \code{\link{df_lvl1_to_lvl2}},
  \code{\link{df_lvl2_folder_structure}},
  \code{\link{df_read_SfnData}},
  \code{\link{df_received_to_accepted}},
  \code{\link{df_report_folder_creation}},
  \code{\link{df_reset_data_status}},
  \code{\link{df_set_status}},
  \code{\link{df_start_status}},
  \code{\link{df_who_ready_to_lvl2}},
  \code{\link{df_write_SfnData}}
}

