% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SfnData_methods.R
\docType{methods}
\name{sfn_get_methods}
\alias{get_env,SfnData-method}
\alias{get_env_flags,SfnData-method}
\alias{get_env_md,SfnData-method}
\alias{get_plant_md,SfnData-method}
\alias{get_sapf,SfnData-method}
\alias{get_sapf_flags,SfnData-method}
\alias{get_si_code,SfnData-method}
\alias{get_site_md,SfnData-method}
\alias{get_species_md,SfnData-method}
\alias{get_stand_md,SfnData-method}
\alias{get_timestamp,SfnData-method}
\alias{sfn_get_methods}
\title{SfnData get methods}
\usage{
\S4method{get_sapf}{SfnData}(object)

\S4method{get_env}{SfnData}(object)

\S4method{get_sapf_flags}{SfnData}(object)

\S4method{get_env_flags}{SfnData}(object)

\S4method{get_timestamp}{SfnData}(object)

\S4method{get_si_code}{SfnData}(object)

\S4method{get_site_md}{SfnData}(object)

\S4method{get_stand_md}{SfnData}(object)

\S4method{get_species_md}{SfnData}(object)

\S4method{get_plant_md}{SfnData}(object)

\S4method{get_env_md}{SfnData}(object)
}
\arguments{
\item{object}{Object of class SfnData from which data is retrieved}

\item{...}{Further arguments to pass on to methods}
}
\description{
Methods to get the info from the SfnData class slots
}

