% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SfnData.R
\docType{methods}
\name{sfn_replacement}
\alias{get_sapf<-,SfnData-method}
\alias{sfn_replacement}
\title{Replacement methods}
\usage{
\S4method{get_sapf}{SfnData}(object) <- value
}
\description{
Methods for replacing the slots with new data or metadata
}

