% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SfnData_methods.R
\docType{methods}
\name{sfn_replacement}
\alias{get_env<-,SfnData-method}
\alias{get_env_flags<-,SfnData-method}
\alias{get_env_md<-,SfnData-method}
\alias{get_plant_md<-,SfnData-method}
\alias{get_sapf<-,SfnData-method}
\alias{get_sapf_flags<-,SfnData-method}
\alias{get_si_code<-,SfnData-method}
\alias{get_site_md<-,SfnData-method}
\alias{get_species_md<-,SfnData-method}
\alias{get_stand_md<-,SfnData-method}
\alias{get_timestamp<-,SfnData-method}
\alias{sfn_replacement}
\title{Sub-setting operation}
\usage{
\S4method{get_sapf}{SfnData}(object) <- value

\S4method{get_env}{SfnData}(object) <- value

\S4method{get_sapf_flags}{SfnData}(object) <- value

\S4method{get_env_flags}{SfnData}(object) <- value

\S4method{get_timestamp}{SfnData}(object) <- value

\S4method{get_si_code}{SfnData}(object) <- value

\S4method{get_site_md}{SfnData}(object) <- value

\S4method{get_stand_md}{SfnData}(object) <- value

\S4method{get_species_md}{SfnData}(object) <- value

\S4method{get_plant_md}{SfnData}(object) <- value

\S4method{get_env_md}{SfnData}(object) <- value
}
\arguments{
\item{object}{SfnData object}

\item{i}{sapflow data row index}

\item{j}{sapflow data column index}

\item{k}{env data row index}

\item{l}{env data column index}
}
\description{
Sub-setting operation
}

