% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SfnData.R
\name{sfn_replacement_generics}
\alias{get_sapf<-}
\alias{sfn_replacement_generics}
\title{Replacement generics}
\usage{
get_sapf(object) <- value
}
\description{
Generic functions for replacement functions for SfnData
}

