% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SfnData_generics.R
\name{sfn_replacement_generics}
\alias{get_env<-}
\alias{get_env_flags<-}
\alias{get_env_md<-}
\alias{get_plant_md<-}
\alias{get_sapf<-}
\alias{get_sapf_flags<-}
\alias{get_si_code<-}
\alias{get_site_md<-}
\alias{get_species_md<-}
\alias{get_stand_md<-}
\alias{get_timestamp<-}
\alias{sfn_replacement_generics}
\title{Replacement generics}
\usage{
get_sapf(object) <- value

get_env(object) <- value

get_sapf_flags(object) <- value

get_env_flags(object) <- value

get_timestamp(object) <- value

get_si_code(object) <- value

get_site_md(object) <- value

get_stand_md(object) <- value

get_species_md(object) <- value

get_plant_md(object) <- value

get_env_md(object) <- value
}
\description{
Generic functions for replacement functions for SfnData
}

