% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_utils.R
\name{sm_status_updater}
\alias{sm_status_updater}
\title{Status updater to server management}
\usage{
sm_status_updater(si_code, parent_logger = "test")
}
\arguments{
\item{si_code}{character with the site code}
}
\value{
Invisible TRUE if no problem was found. Invisible FALSE if the site
  was not updated.
}
\description{
Quick update for status files to update them to the new sapfluxnetQC1 library
versions
}
\details{
This function get the old status, save it, destroy the file, generate a new
one and use the old to update the new one. Use it to generate updated
status files for the sites after updating the QC package
}
\seealso{
Other Server Management: \code{\link{sm_solarTIMESTAMP_adder}}
}
