% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_functions.R
\name{vis_gap_lines}
\alias{vis_gap_lines}
\title{TIMESTAMP with gaps visualization}
\usage{
vis_gap_lines(gaps_info, parent_logger = "test")
}
\arguments{
\item{gaps_info}{Data frame with the gap info as obtained in
\code{\link{qc_mind_the_gap}}}
}
\value{
A ggplot object with the basic lines plot, no themes added.
}
\description{
Concordance lines plot with gaps
}
\seealso{
Other Visualization Functions: \code{\link{vis_plot_the_gap}}
}

