% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_functions.R
\name{vis_gap_lines}
\alias{vis_gap_lines}
\title{TIMESTAMP with gaps visualization}
\usage{
vis_gap_lines(sapf_gaps = NULL, env_gaps = NULL, sapf_intervals = NULL,
  env_intervals = NULL, parent_logger = "test")
}
\arguments{
\item{sapf_gaps}{Data frame with the sapflow gap info as obtained from
\code{\link{qc_mind_the_gap}}}

\item{env_gaps}{Data frame with the env gap info as obtained from
\code{\link{qc_mind_the_gap}}}

\item{sapf_intervals}{Data frame with the sapflow intervals info as obtained from
\code{\link{qc_time_interval}}}

\item{env_intervals}{Data frame with the env intervals info as obtained from
\code{\link{qc_time_interval}}}
}
\value{
A ggplot object with the basic lines plot, no themes added.
}
\description{
Concordance lines plot with gaps
}
\seealso{
Other Visualization Functions: \code{\link{vis_biome}},
  \code{\link{vis_environmental_responses}},
  \code{\link{vis_location_biome}},
  \code{\link{vis_plot_the_gap}}
}
