% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_functions.R
\name{vis_gaps_calendar}
\alias{vis_gaps_calendar}
\title{Gaps density visualization}
\usage{
vis_gaps_calendar(data, parent_logger = "test")
}
\arguments{
\item{data}{Data frame containing the data (env or sf) with the TIMESTAMP
column and the environmental variables or the sapflow measures for the
trees}
}
\value{
Plot is printed and a ggplot2 object is created if there is
  assignation
}
\description{
Heatmap calendar to visualize gaps density
}

